% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2d.R
\name{assign_data}
\alias{assign_data}
\title{Assign data to hexagons}
\usage{
assign_data(data, centroid_df)
}
\arguments{
\item{data}{data A tibble or data frame.}

\item{centroid_df}{The dataset with centroid coordinates.}
}
\value{
A tibble contains embedding components and corresponding hexagon ID.
}
\description{
This function assigns the data to hexagons.
}
\examples{
r2 <- diff(range(s_curve_noise_umap$UMAP2))/diff(range(s_curve_noise_umap$UMAP1))
all_centroids_df <- gen_centroids(bin1 = 3, r2 = r2)
assign_data(data = s_curve_noise_umap_scaled, centroid_df = all_centroids_df)

}
