% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{calc_bins_y}
\alias{calc_bins_y}
\title{Calculate the effective number of bins along x-axis and y-axis}
\usage{
calc_bins_y(nldr_obj, b1 = 4, q = 0.1)
}
\arguments{
\item{b1}{Number of bins along the x axis.}

\item{q}{The buffer amount as proportion of data range.}

\item{nldr_data}{A tibble that contains embedding components in the first and second columns.}
}
\value{
A list of numeric values that represents the effective number of
bins along the y axis, height and, width of the hexagon.
}
\description{
This function calculates the effective number of bins along the x and y axes
of a hexagonal grid.
}
\examples{
calc_bins_y(nldr_obj = scurve_umap_obj, b1 = 4, q = 0.1)

}
