% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocessing.R
\name{calc_y_max}
\alias{calc_y_max}
\title{Compute maximum value of y for scaling}
\usage{
calc_y_max(aspect_ratio, hex_ratio)
}
\arguments{
\item{aspect_ratio}{Numeric value representing the aspect ratio of the plot area.}

\item{hex_ratio}{Numeric value representing the ratio of the hexagon size.}
}
\value{
A value which should be used as maximum value of y when scaling.
}
\description{
This function compute the maximum y value need to use for scaling.
}
\examples{
calc_y_max(aspect_ratio = 2.019414, hex_ratio = 0.2309401)

}
