% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{calculate_effective_shape_value}
\alias{calculate_effective_shape_value}
\title{Calculate effective shape parameter value}
\usage{
calculate_effective_shape_value(.data, x = UMAP1, y = UMAP2)
}
\arguments{
\item{.data}{A data frame, data frame extension (e.g., a tibble), or a lazy data frame (e.g., from dbplyr or dtplyr).}

\item{x}{A literal variable name representing the x-axis.}

\item{y}{A literal variable name representing the y-axis.}
}
\value{
A numeric value representing the effective shape parameter.
}
\description{
This function calculates the effective shape parameter value for a scatter plot based on the ranges of the x-axis and y-axis.
}
\examples{
nldr_df <- s_curve_noise_umap
calculate_effective_shape_value(nldr_df, x = UMAP1, y = UMAP2)
}
