% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{calculate_effective_x_bins}
\alias{calculate_effective_x_bins}
\title{Calculate the effective number of bins along x-axis}
\usage{
calculate_effective_x_bins(nldr_df, x, hex_size = NA, buffer_x = NA)
}
\arguments{
\item{nldr_df}{A data frame containing 2D embeddings.}

\item{x}{The name of the column that contains first 2D embeddings component.}

\item{hex_size}{A numeric value that initializes the radius of the outer circle surrounding the hexagon.}

\item{buffer_x}{The buffer size along the x-axis.}
}
\value{
A numeric value representing the effective number of bins along x-axis.
}
\description{
This function calculates the effective number of bins along the x-axis of a hexagonal grid.
}
\examples{
calculate_effective_x_bins(nldr_df = s_curve_noise_umap_scaled, x = "UMAP1",
hex_size = NA, buffer_x = NA)

}
