% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{calculate_effective_y_bins}
\alias{calculate_effective_y_bins}
\title{Calculate the effective number of bins along y-axis}
\usage{
calculate_effective_y_bins(nldr_df, y, hex_size = NA, buffer_y = NA)
}
\arguments{
\item{nldr_df}{A data frame containing 2D embeddings.}

\item{y}{The name of the column that contains second 2D embeddings component.}

\item{hex_size}{A numeric value that initializes the radius of the outer circle surrounding the hexagon.}

\item{buffer_y}{The buffer size along the y-axis.}
}
\value{
A numeric value representing the effective number of bins along x-axis.
}
\description{
This function calculates the effective number of bins along the y-axis of a hexagonal grid.
}
\examples{
calculate_effective_y_bins(nldr_df = s_curve_noise_umap_scaled, y = "UMAP2",
hex_size = NA, buffer_y = NA)

}
