% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{calculate_opt_bin_val_along_axis}
\alias{calculate_opt_bin_val_along_axis}
\title{Calculate optimal number of bins in a histogram}
\usage{
calculate_opt_bin_val_along_axis(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric value representing the optimal number of bins.
}
\description{
This function calculates the optimal number of bins in a histogram using the IQR method.
}
\examples{
x <- 1:100
calculate_opt_bin_val_along_axis(x)
}
