% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_pca.R
\name{calculate_pca}
\alias{calculate_pca}
\title{Calculate Principal Components}
\usage{
calculate_pca(data, num_pcs)
}
\arguments{
\item{data}{- data set}

\item{num_pcs}{- number of pcs}
}
\description{
Calculate Principal Components
}
