% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2D.R
\name{color_long_edges}
\alias{color_long_edges}
\title{Color Long Edges}
\usage{
color_long_edges(.data, benchmark_value, triangular_object, distance_col)
}
\arguments{
\item{.data}{The data frame containing the edge information.}

\item{benchmark_value}{The threshold value to determine long edges.}

\item{triangular_object}{The triangular object containing the mesh information.}

\item{distance_col}{The column name in `.data` representing the distances.}
}
\value{
A ggplot object with the triangular mesh plot where long edges are colored differently.
}
\description{
This function colors the long edges in a triangular mesh plot based on a benchmark value.
}
\examples{
df <- tibble::tribble(
  ~from, ~to, ~distance,
  1, 2, 5,
  1, 3, 12.2,
  2, 3, 8.25
)
tr_object <- tripack::tri.mesh(df$from, df$to)
color_long_edges(df, 5, tr_object, "distance")

}
