% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2D.R
\name{create_hexbin_df}
\alias{create_hexbin_df}
\title{Generate the Hexagonal Bins}
\usage{
create_hexbin_df(
  .data,
  nldr_df,
  embedding_1,
  embedding_2,
  num_bins = 30,
  shape_val = 1
)
}
\arguments{
\item{.data}{The input data frame containing the original dataset.}

\item{nldr_df}{The non-linear dimensionality reduction data frame containing the 2D coordinates.}

\item{embedding_1}{The name of the column in the non-linear dimensionality reduction data frame containing the first embedding coordinate.}

\item{embedding_2}{The name of the column in the non-linear dimensionality reduction data frame containing the second embedding coordinate.}

\item{num_bins}{The number of hexagonal bins to generate. Default is 30.}

\item{shape_val}{The shape parameter for the hexagonal bins. Default is 1.}
}
\value{
A list with the following elements:
\describe{
  \item{df_new}{A data frame with the merged data and added hexbin IDs.}
  \item{hb}{An S4 object of class "hexbin" representing the generated hexagonal bins.}
}
}
\description{
This function generates hexagonal bins based on the provided data and returns an S4 object of class "hexbin".
}
\examples{
data <- tibble::tribble(
  ~ID, ~x, ~y, ~pc1, ~pc2,
  1, 0.2, 0.5, 0.1, 0.3,
  2, 0.3, 0.4, 0.2, 0.2,
  3, 0.1, 0.7, 0.3, 0.1
)
nldr <- tibble::tribble(
  ~ID, ~embedding1, ~embedding2,
  1, 0.6, 0.9,
  2, 0.7, 0.8,
  3, 0.5, 0.7
)
create_hexbin_df(data, nldr, "embedding1", "embedding2")

}
