% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{extract_coord_of_shifted_hex_grid}
\alias{extract_coord_of_shifted_hex_grid}
\title{Extract Coordinates of Shifted Hexagonal Grid}
\usage{
extract_coord_of_shifted_hex_grid(
  nldr_data_with_hb_id,
  num_bins_x,
  hex_full_count_df,
  shift_x = NA,
  shift_y = NA
)
}
\arguments{
\item{nldr_data_with_hb_id}{A containing 2D embeddings with a hexbin ID.}

\item{num_bins_x}{The number of bins along the x-axis for the hexagonal grid.}

\item{hex_full_count_df}{A data frame with information about all hexagonal grid cells.}

\item{shift_x}{The value that centroids need to be shifted in x-axis. If not provided, it is calculated
as half of the cell diameter of a hexagon.}

\item{shift_y}{The value that centroids need to be shifted in y-axis. If not provided, it is calculated
as half of the cell diameter of a hexagon.}
}
\value{
A data frame with updated hexagon coordinates, hexagon IDs, and counts within each hexagon.

A list containing:
  \item{hex_full_count_df_new}{Data frame with updated hexagonal grid information, including counts and standardized counts.}
  \item{nldr_df_with_new_hexID}{Data frame with new hexagonal bin IDs assigned to 2D embeddings.}
}
\description{
This function takes input data, which includes a hexbin ID, and extracts the coordinates
of a hexagonal grid with a specified shift. The resulting dataset includes hexagon centroids
with updated coordinates and additional information such as counts within each hexagon.
}
\examples{
num_bins_x <- 4
shape_value <- 1.833091
hexbin_data_object <- extract_hexbin_centroids(nldr_df = s_curve_noise_umap,
num_bins = num_bins_x, shape_val = shape_value)
df_bin_centroids <- hexbin_data_object$hexdf_data
hex_full_count_df <- generate_full_grid_info(df_bin_centroids)
UMAP_data_with_hb_id <- s_curve_noise_umap |> dplyr::mutate(hb_id = hexbin_data_object$hb_data@cID)
extract_coord_of_shifted_hex_grid(nldr_data_with_hb_id = UMAP_data_with_hb_id,
num_bins_x = num_bins_x, hex_full_count_df)

}
