% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2D.R
\name{extract_hexbin_centroids}
\alias{extract_hexbin_centroids}
\title{Extract Hexagonal Bin Centroids}
\usage{
extract_hexbin_centroids(.data, hb)
}
\arguments{
\item{.data}{The original data frame containing the data points.}

\item{hb}{An S4 object of class "hexbin" representing the hexagonal bins.}
}
\value{
A data frame containing the extracted centroid information, including the x and y coordinates of each centroid and the corresponding hexbin ID and cell count.
}
\description{
This function computes the x and y coordinates of the centroids from the provided hexagon cell IDs and returns a data frame with the extracted centroid information.
}
\examples{
df <- tibble::tibble(pc1 = c(0.1, 0.2, 0.3), pc2 = c(0.3, 0.2, 0.1), hb_id = c(1, 2, 3))
hb <- hexbin::hexbin(df$pc1, df$pc2, xbins = 30, IDs = TRUE, shape = 1)
extract_hexbin_centroids(df, hb)

}
