% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2D.R
\name{extract_hexbin_mean}
\alias{extract_hexbin_mean}
\title{Extract hexagonal bin mean and related information from non-linear dimensionality reduction data.}
\usage{
extract_hexbin_mean(nldr_df, num_bins, shape_val = 1, x = "UMAP1", y = "UMAP2")
}
\arguments{
\item{nldr_df}{A data frame containing 2D embeddings.}

\item{num_bins}{Number of bins along the x-axis for hexagon binning.}

\item{shape_val}{The value of the shape parameter for hexagon binning.}

\item{x}{The name of the column that contains first 2D embeddings component.}

\item{y}{The name of the column that contains second 2D embeddings component.}
}
\value{
A list containing the hexagonal bin mean data frame and the hexbin object.
}
\description{
Extract hexagonal bin mean and related information from non-linear dimensionality reduction data.
}
\examples{
# Example usage of extract_hexbin_mean function
num_bins_x <- 4
shape_value <- 1.833091
result <- extract_hexbin_mean(nldr_df = s_curve_noise_umap,
num_bins = num_bins_x, shape_val = shape_value, x = "UMAP1", y = "UMAP2")
hexdf_data <- result$hexdf_data
hb_data <- result$hb_data
print(hexdf_data)
print(str(hb_data))


}
