% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2d.R
\name{extract_kmean_centers}
\alias{extract_kmean_centers}
\title{Extract k-mean center coordinates and the corresponding standardize counts.}
\usage{
extract_kmean_centers(nldr_df, tot_bins)
}
\arguments{
\item{nldr_df}{A data frame with 2D embeddings.}

\item{tot_bins}{Total number of bins.}
}
\value{
A tibble contains clustering ID, kmean center coordinates, and standardize counts.
}
\description{
Extract k-mean center coordinates and the corresponding standardize counts.
}
\examples{
num_bins_list <- calc_bins(data = s_curve_noise_umap_scaled, x = "UMAP1", y = "UMAP2",
hex_size = 0.2, buffer_x = 0.346, buffer_y = 0.3)
num_bins_x <- num_bins_list$num_x
num_bins_y <- num_bins_list$num_y
extract_kmean_centers(nldr_df = s_curve_noise_umap |> dplyr::select(-ID),
tot_bins = num_bins_x * num_bins_y)

}
