% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{find_low_density_hexagons}
\alias{find_low_density_hexagons}
\title{Find Low-Density Hexagons}
\usage{
find_low_density_hexagons(df_bin_centroids, num_bins_x, benchmark_rm_hex = NA)
}
\arguments{
\item{df_bin_centroids}{The data frame containing hexagonal bin centroids with density information.}

\item{num_bins_x}{Number of bins along the x-axis for hexagon binning.}

\item{benchmark_rm_hex}{The benchmark mean density below which hexagons will be marked for removal.}
}
\value{
A vector containing the IDs of hexagons to be removed.
}
\description{
This function identifies hexagons with low density based on the mean density of their neighboring hexagons.
}
\examples{
nldr_df <- readRDS(paste0(here::here(), "/quollr/data-raw/s_curve_noise_umap.rds"))
num_bins <- 8
shape_val <- 2.031141
hexbin_data_object <- extract_hexbin_mean(nldr_df, num_bins, shape_val)
df_bin_centroids <- hexbin_data_object$hexdf_data
find_low_density_hexagons(df_bin_centroids, num_bins)


}
