% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2D.R
\name{find_non_empty_bins}
\alias{find_non_empty_bins}
\title{Find the Number of Non-Empty Bins Required}
\usage{
find_non_empty_bins(
  nldr_df,
  x = "UMAP1",
  y = "UMAP2",
  shape_val,
  non_empty_bins
)
}
\arguments{
\item{nldr_df}{Non-linear dimensionality reduction data frame containing 2D coordinates.}

\item{x}{The name of the column that contains first embedding.}

\item{y}{The name of the column that contains second embedding.}

\item{shape_val}{The value of the shape parameter for hexagon binning.}

\item{non_empty_bins}{The desired number of non-empty bins.}
}
\value{
The number of bins along the x-axis needed to achieve the specified number of non-empty bins.
}
\description{
This function determines the number of non-empty bins needed to satisfy a minimum requirement.
}
\examples{
shape_value <- 1.833091
non_empty_bins <- 7
find_non_empty_bins(nldr_df = s_curve_noise_umap, x = "UMAP1", y = "UMAP2",
shape_val = shape_value, non_empty_bins)

}
