% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2d.R
\name{find_pts}
\alias{find_pts}
\title{Find points in hexagonal bins}
\usage{
find_pts(data_hex_id)
}
\arguments{
\item{data_hex_id}{A data frame with data, ID and hexagonal bin IDs.}
}
\value{
A data frame with hexagonal bin IDs and the corresponding points.
}
\description{
This function maps points to their corresponding hexagonal bins.
}
\examples{
num_bins_list <- calc_bins(data = s_curve_noise_umap_scaled, x = "UMAP1", y = "UMAP2",
hex_size = 0.2, buffer_x = 0.346, buffer_y = 0.3)
num_bins_x <- num_bins_list$num_x
num_bins_y <- num_bins_list$num_y
all_centroids_df <- gen_centroids(data = s_curve_noise_umap_scaled,
x = "UMAP1", y = "UMAP2", num_bins_x = num_bins_x,
num_bins_y = num_bins_y, x_start = NA, y_start = NA, buffer_x = 0.346,
buffer_y = 0.3, hex_size = 0.2)
umap_with_hb_id <- assign_data(data = s_curve_noise_umap_scaled,
centroid_df = all_centroids_df, col_start = "UMAP")
find_pts(data_hex_id = umap_with_hb_id)

}
