% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2D.R
\name{find_pts_in_hexbins}
\alias{find_pts_in_hexbins}
\title{Find Points in Hexagonal Bins}
\usage{
find_pts_in_hexbins(full_grid_with_hexbin_id, UMAP_data_with_hb_id)
}
\arguments{
\item{full_grid_with_hexbin_id}{A data frame with hexagonal bin IDs and coordinates.}

\item{UMAP_data_with_hb_id}{A data frame with UMAP data and hexagonal bin IDs.}
}
\value{
A data frame with hexagonal bin IDs and the corresponding points.
}
\description{
This function maps points to their corresponding hexagonal bins based on the provided data frames.
}
\examples{
full_grid_with_hexbin_id <- tibble::tibble(
  hexID = c(1, 2, 3),
  c_x = c(1.5, 2.5, 3.5),
  c_y = c(4.5, 5.5, 6.5)
)
UMAP_data_with_hb_id <- tibble::tibble(
  ID = 101:110,
  UMAP1 = rnorm(10),
  UMAP2 = rnorm(10),
  hb_id = sample(c(1, 2, 3), 10, replace = TRUE)
)
find_pts_in_hexbins(full_grid_with_hexbin_id, UMAP_data_with_hb_id)

}
