% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{fit_highd_model}
\alias{fit_highd_model}
\title{Construct the 2D model and lift into high-D}
\usage{
fit_highd_model(highd_data, nldr_data, b1 = 4, q = 0.1, benchmark_highdens = 5)
}
\arguments{
\item{highd_data}{A tibble that contains the high-dimensional data with a unique identifier.}

\item{nldr_data}{A tibble that contains the embedding with a unique identifier.}

\item{b1}{(default: 4) A numeric value representing the number of bins along the x axis.}

\item{q}{(default: 0.1) A numeric value representing the buffer amount as proportion of data range.}

\item{benchmark_highdens}{(default: 5) A numeric value using to filter high-density hexagons.}
}
\value{
A list containing the data frame with high-dimensional coordinates
for 2D bin centroids (\code{df_bin}) and the data frame containing
information about hexagonal bin centroids (\code{df_bin_centroids}) in 2D.
}
\description{
This function fits a high-dimensional model using hexagonal bins and provides options
to customize the modeling process, including the choice of bin centroids or bin means,
removal of low-density hexagons, and averaging of high-dimensional data.
}
\examples{
fit_highd_model(highd_data = scurve, nldr_data = scurve_umap, b1 = 4,
q = 0.1, benchmark_highdens = 5)

}
