% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2D.R
\name{full_hex_grid}
\alias{full_hex_grid}
\title{Generate Hexagonal Coordinates}
\usage{
full_hex_grid(hexdf_data)
}
\arguments{
\item{hexdf_data}{The dataset with hexbin ID and centroid coordinates.}
}
\value{
A tibble containing the coordinates of hexagons based on hexagonal centroids.
}
\description{
This function generates the coordinates of hexagons after passing hexagonal centroids.
}
\examples{
hexdf_data <- tibble::tibble(
  x = c(0, 1, 0.5),
  y = c(0, 0, 1)
)
full_hex_grid(hexdf_data)

}
