% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{gaussian_clusters}
\alias{gaussian_clusters}
\title{Generate Gaussian Clusters}
\usage{
gaussian_clusters(
  n = 300,
  with_seed = NULL,
  num_clusters = 3,
  mean_matrix = rbind(c(1, 0, 0), c(0, 1, 0), c(0, 0, 1)),
  var_vec = c(0.05, 0.05, 0.05),
  num_dims = 3,
  num_noise_dims = 2,
  min_noise = -0.05,
  max_noise = 0.05
)
}
\arguments{
\item{n}{The number of observations in the dataset.}

\item{with_seed}{Optional. The seed value for reproducibility.}

\item{num_clusters}{The number of Gaussian clusters to generate.}

\item{mean_matrix}{The mean matrix for each cluster, where each row represents the mean vector for a cluster.
Default is rbind(c(1,0,0), c(0,1,0), c(0,0,1)).}

\item{var_vec}{The variance vector for each cluster. Default is c(0.05, 0.05, 0.05).}

\item{num_dims}{The number of dimensions for each observation. Should be at least 2. Default is 3.}

\item{num_noise_dims}{The number of additional random noise dimensions to add to the dataset. Default is 2.}

\item{min_noise}{The minimum value for the random noise dimensions. Default is -0.5.}

\item{max_noise}{The maximum value for the random noise dimensions. Default is 0.5.}
}
\value{
A tibble (data frame) with the generated dataset, including Gaussian clusters and random noise dimensions.
}
\description{
Generates a dataset with Gaussian clusters and additional random noise dimensions.
}
\examples{
data <- gaussian_clusters(n = 300, num_clusters = 3, mean_matrix = rbind(c(1,0,0), c(0,1,0), c(0,0,1)),
                          var_vec = c(0.05, 0.05, 0.05), num_dims = 3, num_noise_dims = 2,
                          min_noise = -0.5, max_noise = 0.5)

}
