% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2d.R
\name{gen_centroids}
\alias{gen_centroids}
\title{Generate centroid coordinate}
\usage{
gen_centroids(bin1 = 2, s1 = -0.1, s2 = -0.1, r2)
}
\arguments{
\item{bin1}{Number of bins along the x axis.}

\item{s1}{The x-coordinate of the hexagonal grid starting point.}

\item{s2}{The y-coordinate of the hexagonal grid starting point.}

\item{r2}{The ratio of the ranges of the original embedding components.}
}
\value{
A tibble contains hexIDs, x and y coordinates (hexID, c_x, c_y respectively)
of all hexagon bin centroids.
}
\description{
This function generates all possible centroids in the hexagonal grid.
}
\examples{
r2 <- diff(range(s_curve_noise_umap$UMAP2))/diff(range(s_curve_noise_umap$UMAP1))
gen_centroids(bin1 = 3, s1 = -0.1, s2 = -0.1, r2 = r2)

}
