% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{gen_design}
\alias{gen_design}
\title{Generate a design to layout 2-D representations}
\usage{
gen_design(n_right, ncol_right = 2)
}
\arguments{
\item{n_right}{The number of plots in right side.}

\item{ncol_right}{The number of columns in right side.}
}
\value{
A patchwork area object.
}
\description{
This function generates a design which can be passed to `plot_layout()`
to arrange 2-D layouts.
}
\examples{
gen_design(n_right = 8, ncol_right = 2)

}
