% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{gen_diffbin1_errors}
\alias{gen_diffbin1_errors}
\title{Generate erros and MSE for different bin widths}
\usage{
gen_diffbin1_errors(highd_data, nldr_data)
}
\arguments{
\item{df_bin_centroids}{Centroid coordinates of hexagonal bins in 2D space.}

\item{df_bin}{Centroid coordinates of hexagonal bins in high dimensions.}

\item{training_data}{Training data used to fit the model.
If NULL, the training data is used (default is NULL).}

\item{type_NLDR}{The type of non-linear dimensionality reduction (NLDR) used.}

\item{col_start}{The text that begin the column name of the high-dimensional data.}
}
\value{
A tibble containing the augmented data with predictions,
error metrics, and absolute error metrics.
}
\description{
This function augments a dataset with predictions and error metrics obtained
from a nonlinear dimension reduction (NLDR) model.
}
\examples{
gen_diffbin1_errors(highd_data = scurve, nldr_data = scurve_umap)

}
