% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2d.R
\name{gen_edges}
\alias{gen_edges}
\title{Generate edge information}
\usage{
gen_edges(tri_object)
}
\arguments{
\item{tri_object}{The triangular object from which to generate edge information.}
}
\value{
A tibble that contains the edge information, including the from-to
relationships and the corresponding x and y coordinates.
}
\description{
This function generates edge information from a given triangular object,
including the coordinates of the vertices and the from-to relationships
between the vertices.
}
\examples{
r2 <- diff(range(s_curve_noise_umap$UMAP2))/diff(range(s_curve_noise_umap$UMAP1))
num_bins_x <- 3
hb_obj <- hex_binning(data = s_curve_noise_umap_scaled, bin1 = num_bins_x,
r2 = r2)
all_centroids_df <- hb_obj$centroids
counts_df <- hb_obj$std_cts
df_bin_centroids <- extract_hexbin_centroids(centroids_df = all_centroids_df,
counts_df = counts_df)
tr1_object <- tri_bin_centroids(hex_df = df_bin_centroids, x = "c_x", y = "c_y")
gen_edges(tri_object = tr1_object)

}
