% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{generate_eval_df}
\alias{generate_eval_df}
\title{Generate Evaluation Data Frame}
\usage{
generate_eval_df(data, prediction_df, df_bin_centroids, df_bin, num_bins)
}
\arguments{
\item{data}{The data set containing high-dimensional data along with IDs.}

\item{prediction_df}{The data set with 2D embeddings, IDs, and predicted hexagonal IDs.}

\item{df_bin_centroids}{The data set with coordinates of hexagonal bin centroids.}

\item{df_bin}{The data set with averaged/weighted high-dimensional data.}

\item{num_bins}{Number of bins along the x-axis for hexagon binning.}
}
\value{
A tibble containing evaluation metrics based on the provided inputs.
}
\description{
This function generates an evaluation data frame based on the provided data and predictions.
}
\examples{
data <- tibble::tibble(
  ID = 1:10,
  x1 = rnorm(10),
  x2 = rnorm(10)
)
prediction_df <- tibble::tibble(
  pred_hb_id = c(1, 2, 1, 3, 2, 3),
  UMAP1 = rnorm(6),
  UMAP2 = rnorm(6),
  ID = seq_along(c(1, 2, 1, 3, 2, 3))  # Adding an ID column to match data
)
df_bin_centroids <- tibble::tibble(
  hexID = 1:3,
  x = rnorm(3),
  y = rnorm(3)
)
df_bin <- tibble::tibble(
  hb_id = c(1, 2, 3),
  avg_x1 = rnorm(3),
  avg_x2 = rnorm(3)
)
num_bins <- 5
prediction_df <- prediction_df |> dplyr::mutate(ID = seq_along(pred_hb_id))
generate_eval_df(data, prediction_df, df_bin_centroids, df_bin, num_bins)

}
