% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2D.R
\name{generate_full_grid_centroids}
\alias{generate_full_grid_centroids}
\title{Generate Full Grid Centroids}
\usage{
generate_full_grid_centroids(hexdf_data)
}
\arguments{
\item{hexdf_data}{The dataset with hexbin ID and centroid coordinates.}
}
\value{
A tibble containing all possible centroids in the full grid.
}
\description{
This function generates all possible centroids in the full grid based on hexbin data.
}
\examples{
num_bins_x <- 4
shape_value <- 1.833091
hexbin_data_object <- extract_hexbin_centroids(nldr_df = s_curve_noise_umap,
num_bins = num_bins_x, shape_val = shape_value)
df_bin_centroids <- hexbin_data_object$hexdf_data
generate_full_grid_centroids(df_bin_centroids)

}
