% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2D.R
\name{generate_full_grid_info}
\alias{generate_full_grid_info}
\title{Generate Full Grid Information Data Frame}
\usage{
generate_full_grid_info(df_bin_centroids)
}
\arguments{
\item{df_bin_centroids}{A data frame with hexagonal bin centroids.}
}
\value{
A data frame with columns "x", "y", "id", "c_x", "c_y", "hexID", "counts", "std_counts", and "polygon_id".
}
\description{
This function generates a data frame containing coordinates and identifiers for hexagons,
along with additional information like counts and polygon IDs.
}
\examples{
nldr_df <- s_curve_noise_umap
num_bins <- 7
shape_val <- 1.833091
set.seed(123)
hexbin_data_object <- extract_hexbin_mean(nldr_df, num_bins, shape_val)
df_bin_centroids <- hexbin_data_object$hexdf_data
df_bin_centroids_sample <- df_bin_centroids |> dplyr::slice_sample(n = 5)
generate_full_grid_info(df_bin_centroids_sample)

}
