% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2d.R
\name{generate_hex_binning_info}
\alias{generate_hex_binning_info}
\title{Generate Hexagonal Binning Info}
\usage{
generate_hex_binning_info(
  nldr_df,
  x = "UMAP1",
  y = "UMAP2",
  num_bins_x,
  num_bins_y,
  x_start = NA,
  y_start = NA,
  buffer_x = NA,
  buffer_y = NA,
  hex_size = NA
)
}
\arguments{
\item{nldr_df}{A data frame containing 2D embeddings.}

\item{x}{The name of the column that contains first 2D embeddings component.}

\item{y}{The name of the column that contains second 2D embeddings component.}

\item{num_bins_x}{Number of bins along the x-axis.}

\item{num_bins_y}{Number of bins along the y-axis.}

\item{x_start}{Starting point along the x-axis for hexagonal binning.}

\item{y_start}{Starting point along the y-axis for hexagonal binning.}

\item{buffer_x}{The buffer size along the x-axis.}

\item{buffer_y}{The buffer size along the y-axis.}

\item{hex_size}{A numeric value that initializes the radius of the outer circle surrounding the hexagon.}
}
\value{
A list contains all hexagonal bin centroids (full_grid_hex_centroids),
hexagonal coordinates of the full grid(full_grid_hex_poly_coordinates),
2D embeddings with corresponding hexagon IDs (nldr_data_with_hex_id),
hex bins with their corresponding standardise counts (hex_id_with_std_counts),
total number of hex bins(total_hex_bins"), number of non-empty hex bins (num_non_empty_bins)
and points within each hexagon (points_in_hex_bins).
}
\description{
This function generates a list which contains hexagonal binning information.
}
\examples{
num_bins_x <- calculate_effective_x_bins(nldr_df = s_curve_noise_umap_scaled,
x = "UMAP1", hex_size = NA, buffer_x = NA)
num_bins_y <- calculate_effective_y_bins(nldr_df = s_curve_noise_umap_scaled,
 y = "UMAP2", hex_size = NA, buffer_y = NA)
generate_hex_binning_info(nldr_df = s_curve_noise_umap_scaled,
x = "UMAP1", y = "UMAP2", num_bins_x = num_bins_x,
num_bins_y = num_bins_y, x_start = NA, y_start = NA, buffer_x = NA,
buffer_y = NA, hex_size = NA)

}
