% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocessing.R
\name{generate_scaled_data}
\alias{generate_scaled_data}
\title{Scaling the 2D embeddings}
\usage{
generate_scaled_data(nldr_df, x = "UMAP1", y = "UMAP2")
}
\arguments{
\item{nldr_df}{A data frame containing 2D embeddings.}

\item{x}{The name of the column that contains first 2D embeddings component.}

\item{y}{The name of the column that contains second 2D embeddings component.}
}
\value{
A list contains scaled 2D embeddings.
}
\description{
This function scales the 2D embeddings.
}
\examples{
generate_scaled_data(nldr_df = s_curve_noise_umap, x = "UMAP1", y = "UMAP2")

}
