% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_hexgrid.R
\docType{data}
\name{GeomHexgrid}
\alias{GeomHexgrid}
\title{GeomHexgrid: A Custom ggplot2 Geom for Hexagonal Grid}
\format{
A ggproto object
}
\usage{
GeomHexgrid
}
\description{
This function defines a custom ggplot2 Geom, GeomHexgrid, for rendering hexagonal grid.
}
\keyword{datasets}
