% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_trimesh.R
\docType{data}
\name{GeomTrimesh}
\alias{GeomTrimesh}
\title{GeomTrimesh: A Custom ggplot2 Geom for Triangular Meshes}
\format{
A ggproto object
}
\usage{
GeomTrimesh
}
\description{
This function defines a custom ggplot2 Geom, GeomTrimesh, for rendering triangular meshes.
}
\details{
- \code{required_aes}: The required aesthetics for this geometry are \code{"x"}, \code{"y"}, \code{"xend"}, and \code{"yend"}.
- \code{default_aes}: The default aesthetics for this geometry include \code{shape = 19}, \code{linetype = 1}, \code{linewidth = 0.5},
  \code{size = 0.5}, \code{alpha = NA}, and \code{colour = "black"}.
- \code{draw_key}: The function describing how to draw the key glyph is \code{ggplot2::draw_key_point}.
- \code{draw_panel}: The function describing how to draw the panel takes \code{data}, \code{panel_scales}, and \code{coord}.
  It creates a tibble of \code{vertices} and a tibble of \code{trimesh}. The final plot is constructed using \code{ggplot2::GeomPoint$draw_panel}
  for vertices and \code{ggplot2::GeomSegment$draw_panel} for trimesh.
}
\keyword{datasets}
