% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2d.R
\name{get_min_indices}
\alias{get_min_indices}
\title{Get indices of all minimum distances}
\usage{
get_min_indices(x)
}
\arguments{
\item{x}{A numeric vector.}
}
\value{
A numeric vector containing the indices of all minimum distances.
}
\description{
This function returns the indices of all minimum distances.
}
\examples{
x <- c(1, 2, 1, 3)
get_min_indices(x)
}
