% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{glance}
\alias{glance}
\title{Generate evaluation metrics}
\usage{
glance(highd_data, model_2d, model_highd)
}
\arguments{
\item{highd_data}{The dataset containing high-dimensional coordinates and an unique identifier.}

\item{model_2d}{Centroid coordinates of hexagonal bins in 2D space.}

\item{model_highd}{Centroid coordinates of hexagonal bins in high dimensions.}
}
\value{
A tibble contains Error, and MSE values.
}
\description{
This function generates an evaluation data frame based on the provided data and predictions.
}
\examples{
df_bin_centroids <- s_curve_obj$s_curve_umap_model_obj$df_bin_centroids
df_bin <- s_curve_obj$s_curve_umap_model_obj$df_bin
glance(highd_data = s_curve_noise_training, model_2d = df_bin_centroids,
model_highd = df_bin)

}
