% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2d.R
\name{hex_binning}
\alias{hex_binning}
\title{Hexagonal binning}
\usage{
hex_binning(nldr_obj, b1 = 4, q = 0.1)
}
\arguments{
\item{b1}{Number of bins along the x axis.}

\item{q}{The buffer amount as proportion of data range.}

\item{nldr_data}{A tibble that contains embedding with a unique identifier.}
}
\value{
A object that contains numeric vector that contains binwidths (a1),
vertical distance (a2), bins along the x and y axes respectively (bins),
numeric vector that contains hexagonal
starting point coordinates all hexagonal bin centroids (centroids),
hexagonal coordinates of the full grid (hex_poly),
embedding components with their corresponding hexagon IDs (data_hb_id),
hex bins with their corresponding standardise counts (std_cts),
total number of hex bins(tot_bins), number of non-empty hex bins (non_bins)
and points within each hexagon (pts_bins).
}
\description{
This function generates the hexagonal object.
}
\examples{
hex_binning(nldr_obj = scurve_umap_obj, b1 = 4, q = 0.1)

}
