% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2D.R
\name{map_hexbin_id}
\alias{map_hexbin_id}
\title{Map HexID to Hexbin Centroids in the Full Grid}
\usage{
map_hexbin_id(full_centroid_df, df_bin_centroids)
}
\arguments{
\item{full_centroid_df}{A data frame containing the full grid centroids.}

\item{df_bin_centroids}{A data frame containing the hexbin centroids.}
}
\value{
A data frame with HexID, centroid coordinates, and standardized counts.
}
\description{
This function maps the HexID to the hexbin centroids in the full grid, creating a data frame that includes the coordinates and standardized counts.
}
\examples{
full_centroid_df <- tibble::tibble(x = c(1, 2, 3), y = c(4, 5, 6))
df_bin_centroids <- tibble::tibble(hexID = c(1, 2, 3), x = c(1.5, 2.5, 3.5), y = c(4.5, 5.5, 6.5), counts = c(10, 15, 5))
map_hexbin_id(full_centroid_df, df_bin_centroids)

}
