% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2D.R
\name{map_polygon_id}
\alias{map_polygon_id}
\title{Map Polygon ID to Hexagon Coordinates}
\usage{
map_polygon_id(full_grid_with_hexbin_id, hex_grid)
}
\arguments{
\item{full_grid_with_hexbin_id}{A data frame containing hexagon IDs, centroids, and standardized counts.}

\item{hex_grid}{A data frame containing all coordinates of hexagons.}
}
\value{
A data frame with hexagon information along with mapped polygon IDs.
}
\description{
This function maps polygon IDs to the corresponding hexagon coordinates in the full grid.
}
\examples{
full_grid_with_hexbin_id <- tibble::tibble(hexID = c(1, 2, 1, 3, 2, 3),
                                           c_x = c(1.5, 2.5, 1.5, 3.5, 2.5, 3.5),
                                           c_y = c(4.5, 5.5, 4.5, 6.5, 5.5, 6.5),
                                           std_counts = c(0.5, 0.75, 0.25, 1, 0.5, 1))
hex_grid <- tibble::tibble(id = c(1, 2, 3), x = c(1, 2, 3), y = c(4, 5, 6))
map_polygon_id(full_grid_with_hexbin_id, hex_grid)

}
