% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{\%>\%}
\alias{\%>\%}
\alias{calculate_effective_number_of_bins}
\alias{calculate_effective_shape_value}
\title{Calculate the effective number of bins along x-axis}
\usage{
calculate_effective_number_of_bins(.data, var1 = var1, var2 = var2)

calculate_effective_shape_value(.data, var1 = var1, var2 = var2)
}
\arguments{
\item{.data}{A data frame, data frame extension (e.g. a tibble), or a lazy data frame (e.g. from dbplyr or dtplyr).}

\item{var1}{A literal variable name which represents x-axis.}

\item{var2}{A literal variable name which represents y-axis.}
}
\value{
A numeric value.

A numeric value.
}
\description{
Calculate the effective number of bins along x-axis

Calculate effective shape parameter value
}
\examples{
data <- tibble::tibble(x = rnorm(300), y = rnorm(300))
calculate_effective_number_of_bins(data, x, y)
data <- tibble::tibble(x = rnorm(300), y = rnorm(300))
calculate_effective_shape_value(data, x, y)
}
