% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{plot_rmse_layouts}
\alias{plot_rmse_layouts}
\title{Arrange RMSE plot and 2-D layouts}
\usage{
plot_rmse_layouts(plots, design)
}
\arguments{
\item{plots}{A list of plots which include RMSE plot and 2-D layouts.}

\item{design}{The design of plots need to be arranged.}
}
\value{
A patchwork object.
}
\description{
This function arranges RMSE plot in left and 2-D layouts in right.
}
\examples{
design <- gen_design(n_right = 4, ncol_right = 2)
plot_rmse_layouts(plots = scurve_plts, design = design)

}
