% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{predict_2d_embeddings}
\alias{predict_2d_embeddings}
\title{Predict 2D Embeddings}
\usage{
predict_2d_embeddings(test_data, df_bin_centroids, df_bin, type_NLDR = "UMAP")
}
\arguments{
\item{test_data}{The test dataset containing high-dimensional features and an ID column.}

\item{df_bin_centroids}{Centroid coordinates of hexagonal bins in 2D space.}

\item{df_bin}{Centroid coordinates of hexagonal bins in high dimensions.}

\item{type_NLDR}{The type of non-linear dimensionality reduction (NLDR) used. Default is "UMAP".}
}
\value{
A data frame with predicted 2D embedding for each data point in the test dataset.
}
\description{
Given a test dataset, the centroid coordinates of hexagonal bins in  2D and high-dimensional space,
predict the 2D embeddings for each data point in the test dataset.
}
\examples{
nldr_df <- s_curve_noise_umap
training_data <- s_curve_noise_training
test_data <- s_curve_noise_test
num_bins <- 8
shape_val <- 2.031141
hexbin_data_object <- extract_hexbin_centroids(nldr_df, num_bins, shape_val)
df_bin_centroids <- hexbin_data_object$hexdf_data
nldr_df_with_hb_id <- nldr_df |>
dplyr::mutate(hb_id = hexbin_data_object_s_curve$hb_data@cID)
df_all <- dplyr::bind_cols(training_data |> dplyr::select(-ID), nldr_df_with_hb_id)
df_bin <- avg_highD_data(.data = df_all, column_start_text = "x")
predict_2d_embeddings(test_data = test_data, df_bin_centroids = df_bin_centroids,
df_bin = df_bin, type_NLDR = "UMAP")

}
