% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{predict_emb}
\alias{predict_emb}
\title{Predict 2D embeddings}
\usage{
predict_emb(highd_data, model_2d, model_highd)
}
\arguments{
\item{highd_data}{The test dataset containing high-dimensional coordinates and an unique identifier.}

\item{model_2d}{Centroid coordinates of hexagonal bins in 2D space.}

\item{model_highd}{Centroid coordinates of hexagonal bins in high dimensions.}
}
\value{
A tibble contains predicted 2D embeddings, ID in the test data, and predicted hexagonal IDs.
}
\description{
Given a test dataset, the centroid coordinates of hexagonal bins in  2D and high-dimensional space,
predict the 2D embeddings for each data point in the test dataset.
}
\examples{
df_bin_centroids <- s_curve_obj$s_curve_umap_model_obj$df_bin_centroids
df_bin <- s_curve_obj$s_curve_umap_model_obj$df_bin
predict_emb(highd_data = s_curve_noise_training, model_2d = df_bin_centroids,
model_highd = df_bin)

}
