% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{predict_hex_id}
\alias{predict_hex_id}
\title{Predict Hexagonal IDs}
\usage{
predict_hex_id(training_data, nldr_df, nldr_df_test, num_bins, shape_val)
}
\arguments{
\item{training_data}{The training dataset containing high-dimensional data with IDs.}

\item{nldr_df}{The non-linear dimensionality reduction data frame containing 2D coordinates of the training data.}

\item{nldr_df_test}{The predicted non-linear dimensionality reductions for the test set.}

\item{num_bins}{Number of bins along the x-axis for hexagon binning.}

\item{shape_val}{The value of the shape parameter for hexagon binning.}
}
\value{
A list containing the predicted data, hexagonal bin centroids, and averaged high-dimensional data.
}
\description{
This function predicts hexagonal IDs for a test set based on training data and non-linear dimensionality reduction (NLDR) coordinates.
}
\examples{
training_data <- tibble::tibble(
  ID = 1:10,
  x1 = rnorm(10),
  x2 = rnorm(10)
)
nldr_df <- tibble::tibble(
  UMAP1 = rnorm(10),
  UMAP2 = rnorm(10)
)
nldr_df_test <- tibble::tibble(
  UMAP1 = rnorm(5),
  UMAP2 = rnorm(5)
)
num_bins <- 5
shape_val <- 1
predict_hex_id(training_data, nldr_df, nldr_df_test, num_bins, shape_val)

}
