% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quollr-package.R
\docType{package}
\name{quollr}
\alias{quollr}
\alias{quollr-package}
\title{quollr: Visualising How Nonlinear Dimension Reduction Warps Your Data}
\description{
\pkg{quollr} helps you assess nonlinear dimension reduction (NLDR) layouts
by constructing a 2-D hexagonal-bin wireframe model and lifting it into
high-dimensional space for visual inspection. This makes it possible to
determine which NLDR method and hyper-parameter choices produce the most
faithful representation of high-dimensional data.
}
\details{
The main workflow is:
\enumerate{
  \item Fit the model with \code{\link{fit_highd_model}}
  \item Combine data and model with \code{\link{comb_data_model}}
  \item Inspect in high dimensions with \code{\link{show_langevitour}}
  \item Evaluate fit with \code{\link{glance}} and \code{\link{augment}}
}
}
\examples{
# Fit the complete model pipeline
model <- fit_highd_model(
  highd_data = scurve,
  nldr_data  = scurve_umap,
  b1         = 15,
  q          = 0.1,
  hd_thresh  = 0
)

# Combine data and model for visualisation
df <- comb_data_model(
  highd_data  = scurve,
  model_highd = model$model_highd,
  model_2d    = model$model_2d
)

# Evaluate goodness of fit
glance(x = model, highd_data = scurve)

# Inspect the lifted model in high dimensions (interactive only)
\dontrun{
  show_langevitour(
    point_data = df,
    edge_data  = model$trimesh_data
  )
}

}
\seealso{
Useful links:
\itemize{
  \item \url{https://jayanilakshika.github.io/quollr/}
  \item \url{https://arxiv.org/abs/2506.22051}
}
}
\author{
\strong{Maintainer}: Jayani P. Gamage \email{jayanilakshika76@gmail.com} (\href{https://orcid.org/0000-0002-6265-6481}{ORCID})

Authors:
\itemize{
  \item Dianne Cook \email{dicook@monash.edu} (\href{https://orcid.org/0000-0002-3813-7155}{ORCID})
  \item Paul Harrison \email{paul.harrison@monash.edu} (\href{https://orcid.org/0000-0002-3980-268X}{ORCID})
  \item Michael Lydeamore \email{michael.lydeamore@monash.edu} (\href{https://orcid.org/0000-0001-6515-827X}{ORCID})
  \item Thiyanga S. Talagala \email{ttalagala@sjp.ac.lk} (\href{https://orcid.org/0000-0002-0656-9789}{ORCID})
}

}
