% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2D.R
\name{remove_long_edges}
\alias{remove_long_edges}
\title{Remove Long Edges from a Triangular Mesh Plot}
\usage{
remove_long_edges(.data, benchmark_value, triangular_object, distance_col)
}
\arguments{
\item{.data}{The data frame containing the edge information.}

\item{benchmark_value}{The threshold value to determine long edges. Edges with a distance greater than or equal to this value will be removed.}

\item{triangular_object}{The triangular object containing the mesh information.}

\item{distance_col}{The column name in `.data` representing the distances.}
}
\value{
A ggplot object with the triangular mesh plot where long edges are removed.
}
\description{
This function removes long edges from a triangular mesh plot based on a benchmark value.
}
\examples{
df <- tibble::tribble(
  ~from, ~to, ~distance,
  1, 2, 5,
  1, 3, 12.2,
  2, 3, 8.25
)
tr_object <- tripack::tri.mesh(df$from, df$to)
remove_long_edges(df, 10, tr_object, "distance")

}
