% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{s_curve_noise_umap}
\alias{s_curve_noise_umap}
\title{UMAP embedding for S-curve dataset which with noise dimensions}
\format{
## `s_curve_noise_umap`
A data frame with 225 rows and 3 columns:
\describe{
  \item{UMAP1}{Numeric, UMAP coordinate 1.}
  \item{UMAP2}{Numeric, UMAP coordinate 2.}
  \item{ID}{Numeric, identifier for each data point.}
}
}
\source{
This dataset is generated for illustrative purposes.
}
\usage{
data(s_curve_noise_umap)
}
\description{
The `s_curve_noise_umap` dataset contains the UMAP (Uniform Manifold
Approximation and Projection) embeddings of a three-dimensional S-curve with
added noise. Each data point is represented by two UMAP coordinates (UMAP1
and UMAP2) and an ID.
}
\examples{
# Load the s_curve_noise_umap dataset
data(s_curve_noise_umap)

# Display the first few rows of the dataset
head(s_curve_noise_umap)

}
\keyword{datasets}
