% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{s_curve_noise_umap2}
\alias{s_curve_noise_umap2}
\title{UMAP embedding for S-curve dataset which with noise dimensions}
\format{
## `s_curve_noise_umap2`
A data frame with 75 rows and 3 columns:
\describe{
  \item{UMAP1}{Numeric, first UMAP 2D embeddings.}
  \item{UMAP2}{Numeric, second UMAP 2D embeddings.}
  \item{ID}{Numeric, identifier for each data point.}
}
}
\source{
This dataset is generated for illustrative purposes.
}
\usage{
data(s_curve_noise_umap2)
}
\description{
The `s_curve_noise_umap2` dataset contains the UMAP (Uniform Manifold
Approximation and Projection) embeddings of a three-dimensional S-curve with
added noise. Each data point is represented by two UMAP coordinates (UMAP1
and UMAP2) and an ID.
}
\examples{
# Load the s_curve_noise_umap2 dataset
data(s_curve_noise_umap2)

# Display the first few rows of the dataset
head(s_curve_noise_umap2)

}
\keyword{datasets}
