% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{s_curve_noise_umap_scaled}
\alias{s_curve_noise_umap_scaled}
\title{Scaled UMAP embedding for S-curve dataset which with noise dimensions}
\format{
## `s_curve_noise_umap_scaled`
A data frame with 25 rows and 3 columns:
\describe{
  \item{UMAP1}{Numeric, Scaled first UMAP 2D embeddings.}
  \item{UMAP2}{Numeric, Scaled second UMAP 2D embedding.}
  \item{ID}{Numeric, identifier for each data point.}
}
}
\source{
This dataset is generated for illustrative purposes.
}
\usage{
data(s_curve_noise_umap_scaled)
}
\description{
The `s_curve_noise_umap_scaled` dataset contains the scaled UMAP (Uniform Manifold
Approximation and Projection) embeddings.
}
\examples{
# Load the s_curve_noise_umap_scaled dataset
data(s_curve_noise_umap_scaled)

# Display the first few rows of the dataset
head(s_curve_noise_umap_scaled)

}
\keyword{datasets}
