% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{model object}
\name{s_curve_obj}
\alias{s_curve_obj}
\title{Object for S-curve dataset}
\format{
## `s_curve_obj`
An object of 6 elements
}
\source{
This object is generated for illustrative purposes.
}
\usage{
data(s_curve_obj)
}
\description{
The `s_curve_obj` contains object of scaled umap embedding, x and y limits,
range of y; object of hexagonal binning information,
object of high-d model fitted to umap embedding for the training data,
the triangular object representing the triangulated bin centroids,
a tibble that contains the edge information, a tibble with
edge distance information.
}
\examples{
# Load the s_curve_obj
data(s_curve_obj)

}
\keyword{object}
