% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{s_curve_with_noise_dims}
\alias{s_curve_with_noise_dims}
\title{Generate S-curve Data with Additional Noise Dimensions}
\usage{
s_curve_with_noise_dims(
  n = 200,
  with_seed = NULL,
  num_noise_dims = 3,
  min_noise = -0.5,
  max_noise = 0.5
)
}
\arguments{
\item{n}{The number of observations in the dataset.}

\item{with_seed}{Optional. The seed value for reproducibility.}

\item{num_noise_dims}{The number of additional random noise dimensions to add to the dataset. Default is 8.}

\item{min_noise}{The minimum value for the random noise dimensions. Default is -0.5.}

\item{max_noise}{The maximum value for the random noise dimensions. Default is 0.5.}
}
\value{
A tibble (data frame) with the generated dataset, including an S-curve and random noise dimensions.
}
\description{
Generates an S-curve dataset with additional random noise dimensions.
}
\examples{
data <- s_curve_with_noise_dims(n = 200, num_noise_dims = 3,
                                min_noise = -0.5, max_noise = 0.5)

}
