% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scurve_plts}
\alias{scurve_plts}
\title{List of plots}
\format{
## `scurve_plts`
A list of 5 elements:
\describe{
  \item{scurve_plts[[1]]}{ggplot object, RMSE plot.}
  \item{scurve_plts[[2]]}{ggplot object, 2-D NLDR layout for `scurve_umap`.}
  \item{scurve_plts[[3]]}{ggplot object, 2-D NLDR layout for `scurve_umap2`.}
  \item{scurve_plts[[4]]}{ggplot object, 2-D NLDR layout for `scurve_umap3`.}
  \item{scurve_plts[[5]]}{ggplot object, 2-D NLDR layout for `scurve_umap4`.}
}
}
\source{
This list of plots is generated for illustrative purposes.
}
\usage{
scurve_plts
}
\description{
The `scurve_plts` contains the RMSE plot and the 2-D NLDR layouts for
`scurve_umap`, `scurve_umap2`, `scurve_umap3`, and `scurve_umap4`
}
\examples{
# Show the scurve_plts
scurve_plts

}
\keyword{plots}
