% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scurve_umap4}
\alias{scurve_umap4}
\title{UMAP embedding for `scurve` with n_neighbors = 30 and min_dist = 0.5}
\format{
## `scurve_umap4`
A data frame with 5000 rows and 3 columns:
\describe{
  \item{UMAP1}{Numeric, first UMAP 2-D embeddings.}
  \item{UMAP2}{Numeric, second UMAP 2-D embeddings.}
  \item{ID}{Numeric, identifier for each data point.}
}
}
\source{
This dataset is generated for illustrative purposes.
}
\usage{
data(scurve_umap4)
}
\description{
The `scurve_umap4` dataset contains the UMAP (Uniform Manifold
Approximation and Projection) embeddings of a three-dimensional S-curve with
added noise. Each data point is represented by two UMAP coordinates (UMAP1
and UMAP2) and an ID.
}
\examples{
# Load the scurve_umap4 dataset
data(scurve_umap4)

# Display the first few rows of the dataset
head(scurve_umap4)

}
\keyword{datasets}
