% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scurve_umap5}
\alias{scurve_umap5}
\title{UMAP embedding for `scurve` with n_neighbors = 15 and min_dist = 0.5}
\format{
## `scurve_umap5`
A data frame with 5000 rows and 3 columns:
\describe{
  \item{UMAP1}{Numeric, first UMAP 2-D embeddings.}
  \item{UMAP2}{Numeric, second UMAP 2-D embeddings.}
  \item{ID}{Numeric, identifier for each data point.}
}
}
\source{
This dataset is generated for illustrative purposes.
}
\usage{
data(scurve_umap5)
}
\description{
The `scurve_umap5` dataset contains the UMAP (Uniform Manifold
Approximation and Projection) embeddings of a three-dimensional S-curve with
added noise. Each data point is represented by two UMAP coordinates (UMAP1
and UMAP2) and an ID.
}
\examples{
# Load the scurve_umap5 dataset
data(scurve_umap5)

# Display the first few rows of the dataset
head(scurve_umap5)

}
\keyword{datasets}
