% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scurve_umap_mse5}
\alias{scurve_umap_mse5}
\title{UMAP embedding for S-curve dataset which with noise dimensions}
\format{
## `scurve_umap_mse5`
A data frame with 75 rows and 3 columns:
\describe{
  \item{UMAP1}{Numeric, first UMAP 2D embeddings.}
  \item{UMAP2}{Numeric, second UMAP 2D embeddings.}
  \item{ID}{Numeric, identifier for each data point.}
}
}
\source{
This dataset is generated for illustrative purposes.
}
\usage{
data(scurve_umap_mse5)
}
\description{
The `scurve_umap_mse5` dataset contains the UMAP (Uniform Manifold
Approximation and Projection) embeddings of a three-dimensional S-curve with
added noise. Each data point is represented by two UMAP coordinates (UMAP1
and UMAP2) and an ID.
}
\examples{
# Load the scurve_umap_mse5 dataset
data(scurve_umap_mse5)

# Display the first few rows of the dataset
head(scurve_umap_mse5)

}
\keyword{datasets}
