% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scurve_umap_obj}
\alias{scurve_umap_obj}
\title{Scaled UMAP embedding for S-curve dataset which with noise dimensions}
\format{
## `scurve_umap_obj`
A data frame with 25 rows and 3 columns:
\describe{
  \item{UMAP1}{Numeric, Scaled first UMAP 2D embeddings.}
  \item{UMAP2}{Numeric, Scaled second UMAP 2D embedding.}
  \item{ID}{Numeric, identifier for each data point.}
}
}
\source{
This dataset is generated for illustrative purposes.
}
\usage{
data(scurve_umap_obj)
}
\description{
The `scurve_umap_obj` dataset contains the scaled UMAP (Uniform Manifold
Approximation and Projection) embeddings.
}
\examples{
# Load the scurve_umap_obj dataset
data(scurve_umap_obj)

# Display the first few rows of the dataset
head(scurve_umap_obj)

}
\keyword{datasets}
