% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scurve_umap_rmse2}
\alias{scurve_umap_rmse2}
\title{Summary with different number of bins for `scurve_umap2`}
\format{
## `scurve_umap_rmse2`
A data frame with 70 rows and 10 columns:
\describe{
  \item{Error}{Numeric, model error.}
  \item{RMSE}{Numeric, Root Mean Square Error.}
  \item{b1}{Numeric, number of bins along x-axis.}
  \item{b2}{Numeric, number of bins along y-axis.}
  \item{b}{Numeric, number of total bins.}
  \item{m}{Numeric, number of non-empty bins.}
  \item{a1}{Numeric, hexagon bin width.}
  \item{a2}{Numeric, hexagon bin height.}
  \item{d_bar}{Numeric, mean desnity.}
  \item{method}{Character, NLDR method.}
}
}
\source{
This dataset is generated for illustrative purposes.
}
\usage{
data(scurve_umap_rmse2)
}
\description{
The `scurve_umap_rmse2` dataset contains error, RMSE, b2, b, m, a1, a2,
and mean density (d_bar) for different number of bins in x-axis (b1).
}
\examples{
# Load the scurve_umap_rmse2 dataset
data(scurve_umap_rmse2)

# Display the first few rows of the dataset
head(scurve_umap_rmse2)

}
\keyword{datasets}
