% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_data_link_plots.R
\name{show_error_link_plots}
\alias{show_error_link_plots}
\title{Visualize the model overlaid on high-dimensional data along with 2D wireframe model and error.}
\usage{
show_error_link_plots(point_df, edge_df)
}
\arguments{
\item{point_df}{A tibble that contains the high-dimensional data, no-linear dimension reductions
and model in high-dimensions.}

\item{edge_df}{A tibble that contains the wireframe data (from and to).}

\item{error_df}{A tibble that high-dimesional model error.}
}
\value{
A browsable HTML element.
}
\description{
This function generates a LangeviTour visualization based on different
conditions and input parameters with 2D wireframe.
}
\examples{
df_bin_centroids <- s_curve_obj$s_curve_umap_model_obj$df_bin_centroids
df_bin <- s_curve_obj$s_curve_umap_model_obj$df_bin
model_error <- augment(model_2d = df_bin_centroids, model_highd = df_bin,
highd_data = s_curve_noise_training)
df_exe <- comb_all_data_model_error(highd_data = s_curve_noise_training,
nldr_data = s_curve_obj$s_curve_umap_scaled_obj$scaled_nldr,
model_highd = df_bin, model_2d = df_bin_centroids, error_df = model_error)
edge_data <- s_curve_obj$s_curve_umap_model_tr_from_to_df
show_error_link_plots(point_df = df_exe, edge_df = edge_data)

}
