% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_hexgrid.R
\name{stat_hexgrid}
\alias{stat_hexgrid}
\title{stat_hexgrid Custom Stat for hexagonal grid plot}
\usage{
stat_hexgrid(
  mapping = NULL,
  data = NULL,
  geom = GeomHexgrid$default_aes(),
  position = "identity",
  show.legend = NA,
  outliers = TRUE,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{mapping}{Aesthetic mappings for the plot.}

\item{data}{The data to be plotted.}

\item{geom}{The geometry to be used in the plot.}

\item{position}{The position adjustment to be applied.}

\item{show.legend}{Whether to show the legend for this layer.}

\item{outliers}{Whether to include outliers.}

\item{inherit.aes}{Whether to inherit aesthetics from the plot or the layer.}

\item{...}{Additional arguments to be passed to the `layer` function.}
}
\value{
A `ggplot2` layer object.
}
\description{
stat_hexgrid Custom Stat for hexagonal grid plot
}
