% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2d.R
\name{tri_bin_centroids}
\alias{tri_bin_centroids}
\title{Triangulate bin centroids}
\usage{
tri_bin_centroids(hex_df, x, y)
}
\arguments{
\item{hex_df}{The tibble containing the bin centroids.}

\item{x}{The name of the column that contains x coordinates of bin centroids.}

\item{y}{The name of the column that contains y coordinates of bin centroids.}
}
\value{
A triangular object representing the triangulated bin centroids.
}
\description{
This function triangulates the bin centroids using the x and y coordinates
provided in the input data frame and returns the triangular object.
}
\examples{
r2 <- diff(range(s_curve_noise_umap$UMAP2))/diff(range(s_curve_noise_umap$UMAP1))
num_bins_x <- 3
hb_obj <- hex_binning(data = s_curve_noise_umap_scaled, bin1 = num_bins_x,
q = 0.1, r2 = r2)
all_centroids_df <- hb_obj$centroids
counts_df <- hb_obj$std_cts
df_bin_centroids <- extract_hexbin_centroids(centroids_df = all_centroids_df,
counts_df = counts_df)
tri_bin_centroids(hex_df = df_bin_centroids, x = "c_x", y = "c_y")

}
