% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{triangular_plane_with_bkg_noise}
\alias{triangular_plane_with_bkg_noise}
\title{Generate Data with Triangular Plane and Background Noise}
\usage{
triangular_plane_with_bkg_noise(
  sample_size = 675,
  with_seed = NULL,
  num_of_noise_dim = 4,
  min_noise = -0.5,
  max_noise = 0.5
)
}
\arguments{
\item{sample_size}{Number of total data points.}

\item{with_seed}{Seed value for reproducibility. If NULL, no seed is set.}

\item{num_of_noise_dim}{Number of additional noise dimensions to be added to the data.}

\item{min_noise}{Minimum value for generating noise.}

\item{max_noise}{Maximum value for generating noise.}
}
\value{
A dataset with synthetic data points and background noise.
}
\description{
Generate synthetic data points with a triangular plane and background noise. This function
creates a dataset with a triangular cluster in two dimensions and adds background noise to
the points. The number of data points and the number of noise dimensions can be controlled.
}
\examples{
data <- triangular_plane_with_bkg_noise(sample_size = 675, with_seed = 42, num_of_noise_dim = 4,
                                        min_noise = -0.5, max_noise = 0.5)

}
