% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2D.R
\name{triangulate_bin_centroids}
\alias{triangulate_bin_centroids}
\title{Triangulate Bin Centroids}
\usage{
triangulate_bin_centroids(.data)
}
\arguments{
\item{.data}{The data frame containing the bin centroids.}
}
\value{
A triangular object representing the triangulated bin centroids.
}
\description{
This function triangulates the bin centroids using the x and y coordinates provided in the input data frame and returns the triangular object.
}
\examples{
df <- tibble::tibble(x_val_center = rnorm(100), y_val_center = rnorm(100))
triangulate_bin_centroids(df)

}
