% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2d.R
\name{update_trimesh_index}
\alias{update_trimesh_index}
\title{Update from and to values in trimesh data}
\usage{
update_trimesh_index(trimesh_data)
}
\arguments{
\item{trimesh_data}{A tibble that contains wireframe data.}
}
\value{
A tibble that contains the updated edge information.
}
\description{
This function update the from and to indexes.
}
\examples{
tr_from_to_df <- scurve_model_obj$trimesh_data
update_trimesh_index(trimesh_data = tr_from_to_df)

}
