% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_2d.R
\name{vis_lg_mesh}
\alias{vis_lg_mesh}
\title{Visualize triangular mesh with coloured long edges}
\usage{
vis_lg_mesh(distance_edges, benchmark_value, tr_coord_df, distance_col)
}
\arguments{
\item{distance_edges}{The tibble that contains the edge information.}

\item{benchmark_value}{The threshold value to determine long edges.}

\item{tr_coord_df}{A tibble that contains the x and y coordinates of start and end points.}

\item{distance_col}{The column name in `distance_edges` representing the distances.}
}
\value{
A ggplot object with the triangular mesh plot where long edges are
coloured differently.
}
\description{
This function visualize triangular mesh with coloured long edges.
}
\examples{
tr_from_to_df <- s_curve_obj$s_curve_umap_model_tr_from_to_df
distance_df <- s_curve_obj$distance_df
vis_lg_mesh(distance_edges = distance_df, benchmark_value = 0.75,
tr_coord_df = tr_from_to_df, distance_col = "distance")

}
