% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2d.R
\name{vis_mesh}
\alias{vis_mesh}
\title{Visualize triangular mesh after removing the long edges}
\usage{
vis_mesh(trimesh_data)
}
\arguments{
\item{trimesh_data}{A tibble that contains the x and y coordinates of start and end points.}
}
\value{
A ggplot object with the triangular mesh plot where long edges are removed.
}
\description{
This function visualize the triangular mesh after removing the long edges.
}
\examples{
tr_from_to_df <- scurve_model_obj$trimesh_data
vis_mesh(trimesh_data = tr_from_to_df)

}
