% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2d.R
\name{vis_rmlg_mesh}
\alias{vis_rmlg_mesh}
\title{Visualize triangular mesh after removing the long edges}
\usage{
vis_rmlg_mesh(tr_coord_df)
}
\arguments{
\item{tr_coord_df}{A tibble that contains the x and y coordinates of start and end points.}
}
\value{
A ggplot object with the triangular mesh plot where long edges are removed.
}
\description{
This function visualize the triangular mesh after removing the long edges.
}
\examples{
tr_from_to_df <- s_curve_obj$s_curve_umap_model_tr_from_to_df
vis_rmlg_mesh(tr_coord_df = tr_from_to_df)

}
