% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highD.R
\name{weighted_highD_data}
\alias{weighted_highD_data}
\title{Compute Weighted Mean for High-Dimensional Data}
\usage{
weighted_highD_data(
  training_data,
  nldr_df_with_id,
  hb_object,
  column_start_text = "x"
)
}
\arguments{
\item{training_data}{A data frame containing the training data with an ID column.}

\item{nldr_df_with_id}{A data frame containing the non-linear dimensionality reduction (NLDR) data with an ID column.}

\item{hb_object}{An object containing information about hexbin IDs.}

\item{column_start_text}{The starting text of the column names in the training_data that should be considered for the weighted mean. Default is "x".}
}
\value{
A data frame with the computed weighted mean for each specified column.
}
\description{
This function computes the weighted mean of the specified columns in the training data
based on the distances from the average points in the non-linear dimensionality reduction (NLDR) space.
}
\examples{
training_data <- s_curve_noise_training
nldr_df_with_id <- s_curve_noise_umap
num_bins <- 8
shape_val <- 2.031141
result <- extract_hexbin_centroids(nldr_df, num_bins, shape_val)
hexdf_data <- result$hexdf_data
hb_object <- result$hb_data
weighted_highD_data(training_data, nldr_df_with_id, hb_object)

}
\seealso{
\code{\link{compute_weights}}
}
