% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2d.R
\name{assign_data}
\alias{assign_data}
\title{Assign data to hexagons}
\usage{
assign_data(nldr_obj, centroids_data)
}
\arguments{
\item{centroids_data}{The dataset with centroid coordinates.}

\item{nldr_data}{A tibble that contains embedding components in the first and second columns.}
}
\value{
A tibble contains embedding components and corresponding hexagon ID.
}
\description{
This function assigns the data to hexagons.
}
\examples{
all_centroids_df <- scurve_model_obj$hb_obj$centroids
assign_data(nldr_obj = scurve_umap_obj, centroids_data = all_centroids_df)

}
