% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{augment}
\alias{augment}
\title{Augment Data with Predictions and Error Metrics}
\usage{
augment(highd_data, model_2d, model_highd)
}
\arguments{
\item{df_bin_centroids}{Centroid coordinates of hexagonal bins in 2D space.}

\item{df_bin}{Centroid coordinates of hexagonal bins in high dimensions.}

\item{training_data}{Training data used to fit the model.
If NULL, the training data is used (default is NULL).}

\item{type_NLDR}{The type of non-linear dimensionality reduction (NLDR) used.}

\item{col_start}{The text that begin the column name of the high-dimensional data.}
}
\value{
A tibble containing the augmented data with predictions,
error metrics, and absolute error metrics.
}
\description{
This function augments a dataset with predictions and error metrics obtained
from a nonlinear dimension reduction (NLDR) model.
}
\examples{
df_bin_centroids <- s_curve_obj$s_curve_umap_model_obj$df_bin_centroids
df_bin <- s_curve_obj$s_curve_umap_model_obj$df_bin
augment(highd_data = s_curve_noise_training, model_2d = df_bin_centroids,
model_highd = df_bin)

}
