% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{augment}
\alias{augment}
\title{S3 generic for augment}
\usage{
augment(x, ...)
}
\arguments{
\item{x}{An object to augment.}

\item{...}{Additional arguments passed to methods (e.g., high-dimensional data).}
}
\description{
This is a generic function to augment datasets with predictions and error metrics
from different models, including NLDR models.
}
