% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highd.R
\name{avg_highd_data}
\alias{avg_highd_data}
\title{Create a dataframe with averaged high-dimensional data}
\usage{
avg_highd_data(data)
}
\arguments{
\item{data}{A tibble that contains the high-dimensional data and embedding
with hexagonal bin IDs.}
}
\value{
A tibble with the average values of the high-dimensional data within each hexagonal bin.
}
\description{
This function calculates the average values of high-dimensional data within each hexagonal bin.
}
\examples{
umap_with_hb_id <- s_curve_obj$s_curve_umap_hb_obj$data_hb_id
df_all <- dplyr::bind_cols(s_curve_noise_training, umap_data_with_hb_id)
avg_highd_data(data = df_all)

}
