% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highD.R
\name{avg_highD_data}
\alias{avg_highD_data}
\title{Average High-Dimensional Data}
\usage{
avg_highD_data(.data, column_start_text = "x")
}
\arguments{
\item{.data}{The data frame containing the high-dimensional data.}

\item{column_start_text}{The text that begin the column name of the high-D data}
}
\value{
A data frame with the average values of the high-dimensional data within each hexagonal bin.
}
\description{
This function calculates the average values of high-dimensional data within each hexagonal bin.
}
\examples{
nldr_df <- s_curve_noise_umap
training_data <- s_curve_noise_training
num_bins <- 8
shape_val <- 2.031141
hexbin_data_object <- extract_hexbin_mean(nldr_df, num_bins, shape_val)
df_bin_centroids <- hexbin_data_object$hexdf_data
UMAP_data_with_hb_id <- nldr_df |> dplyr::mutate(hb_id = hexbin_data_object$hb_data@cID)
df_all <- dplyr::bind_cols(training_data |> dplyr::select(-ID), UMAP_data_with_hb_id)
avg_highD_data(df_all)

}
