% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highD.R
\name{avg_highD_data}
\alias{avg_highD_data}
\title{Average High-Dimensional Data}
\usage{
avg_highD_data(.data, column_start_text = "x")
}
\arguments{
\item{.data}{The data frame containing the high-dimensional data.}
}
\value{
A data frame with the average values of the high-dimensional data within each hexagonal bin.
}
\description{
This function calculates the average values of high-dimensional data within each hexagonal bin.
}
\examples{
df <- tibble::tribble(
  ~x1, ~x2, ~x3, ~hb_id,
  1, 2, 3, 1,
  4, 5, 6, 2,
  7, 8, 9, 1
)
avg_highD_data(df)

}
