% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highd.R
\name{avg_highD_data}
\alias{avg_highD_data}
\title{Average High-Dimensional Data}
\usage{
avg_highD_data(.data, column_start_text = "x")
}
\arguments{
\item{.data}{A data frame containing the high-dimensional data and 2D embeddings with hexagonal bin IDs.}

\item{column_start_text}{The text that begin the column name of the high-dimensional data}
}
\value{
A data frame with the average values of the high-dimensional data within each hexagonal bin.
}
\description{
This function calculates the average values of high-dimensional data within each hexagonal bin.
}
\examples{
training_data <- s_curve_noise_training
num_bins_x <- calculate_effective_x_bins(nldr_df = s_curve_noise_umap_scaled,
x = "UMAP1", hex_size = NA, buffer_x = NA)
num_bins_y <- calculate_effective_y_bins(nldr_df = s_curve_noise_umap_scaled,
 y = "UMAP2", hex_size = NA, buffer_y = NA)
hex_bin_obj <- generate_hex_binning_info(nldr_df = s_curve_noise_umap_scaled,
x = "UMAP1", y = "UMAP2", num_bins_x = num_bins_x,
num_bins_y = num_bins_y, x_start = NA, y_start = NA, buffer_x = NA,
buffer_y = NA, hex_size = NA)
UMAP_data_with_hb_id <- hex_bin_obj$nldr_data_with_hex_id
df_all <- dplyr::bind_cols(training_data |> dplyr::select(-ID), UMAP_data_with_hb_id)
avg_highD_data(df_all, column_start_text = "x")

}
