% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highd.R
\name{avg_highd_data}
\alias{avg_highd_data}
\title{Create a dataframe with averaged high-dimensional data}
\usage{
avg_highd_data(data, col_start = "x")
}
\arguments{
\item{data}{A tibble that contains the high-dimensional data and embedding
with hexagonal bin IDs.}

\item{col_start}{The text that begin the column name of the high-dimensional data}
}
\value{
A tibble with the average values of the high-dimensional data within each hexagonal bin.
}
\description{
This function calculates the average values of high-dimensional data within each hexagonal bin.
}
\examples{
r2 <- diff(range(s_curve_noise_umap$UMAP2))/diff(range(s_curve_noise_umap$UMAP1))
num_bins_x <- 3
hb_obj <- hex_binning(data = s_curve_noise_umap_scaled, bin1 = num_bins_x,
r2 = r2)
umap_data_with_hb_id <- hb_obj$data_hb_id
df_all <- dplyr::bind_cols(s_curve_noise_training, umap_data_with_hb_id)
avg_highd_data(data = df_all, col_start = "x")

}
