% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highd.R
\name{avg_highd_data}
\alias{avg_highd_data}
\title{Create a tibble with averaged high-dimensional data}
\usage{
avg_highd_data(highd_data, scaled_nldr_hexid)
}
\arguments{
\item{highd_data}{A tibble that contains the high-dimensional data.}

\item{scaled_nldr_hexid}{A tibble that contains the scaled embedding with hexagonal bin IDs.}
}
\value{
A tibble with the average values of the high-dimensional data within each hexagonal bin.
}
\description{
This function calculates the average values of high-dimensional data within each hexagonal bin.
}
\examples{
umap_with_hb_id <- scurve_model_obj$hb_obj$data_hb_id
avg_highd_data(highd_data = scurve, scaled_nldr_hexid = umap_with_hb_id)

}
