% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2D.R
\name{cal_2D_dist}
\alias{cal_2D_dist}
\title{Calculate 2D Distances}
\usage{
cal_2D_dist(.data)
}
\arguments{
\item{.data}{The data frame containing the points for which to calculate distances.}
}
\value{
A data frame with the from-to relationships and the corresponding 2D distances.
}
\description{
This function calculates the 2D distances between points in a given data frame.
}
\examples{
nldr_df <- readRDS(paste0(here::here(), "/quollr/data-raw/s_curve_noise_umap.rds"))
num_bins <- 8
shape_val <- 2.031141
hexbin_data_object <- extract_hexbin_mean(nldr_df, num_bins, shape_val)
df_bin_centroids <- hexbin_data_object$hexdf_data
tr1_object <- triangulate_bin_centroids(df_bin_centroids, x, y)
tr_from_to_df <- generate_edge_info(triangular_object = tr1_object)
cal_2D_dist(tr_from_to_df)

}
