% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2D.R
\name{cal_2D_dist}
\alias{cal_2D_dist}
\title{Calculate 2D Distances}
\usage{
cal_2D_dist(.data)
}
\arguments{
\item{.data}{The data frame containing the points for which to calculate distances.}
}
\value{
A data frame with the from-to relationships and the corresponding 2D distances.
}
\description{
This function calculates the 2D distances between points in a given data frame.
}
\examples{
df <- tibble::tribble(
  ~from, ~to, ~x_from, ~y_from, ~x_to, ~y_to,
  1, 2, 6, 0, 3, 4,
  1, 3, 7, 0, 5, 12,
  2, 3, 3, 4, 5, 12
)
cal_2D_dist(df)

}
