% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2d.R
\name{cal_2d_dist}
\alias{cal_2d_dist}
\title{Calculate 2D Euclidean distances between vertices}
\usage{
cal_2d_dist(tr_coord_df, start_x, start_y, end_x, end_y, select_vars)
}
\arguments{
\item{tr_coord_df}{A tibble that contains the x and y coordinates of start
and end points.}

\item{start_x}{Column name for the x-coordinate of the starting point.}

\item{start_y}{Column name for the y-coordinate of the starting point.}

\item{end_x}{Column name for the x-coordinate of the ending point.}

\item{end_y}{Column name for the y-coordinate of the ending point.}

\item{select_vars}{A character vector specifying the columns to be
selected in the resulting data frame.}
}
\value{
A tibble with columns for the starting point, ending point,
and calculated distances.
}
\description{
This function calculates the 2D distances between pairs of points in a data frame.
}
\examples{
r2 <- diff(range(s_curve_noise_umap$UMAP2))/diff(range(s_curve_noise_umap$UMAP1))
num_bins_x <- 3
hb_obj <- hex_binning(data = s_curve_noise_umap_scaled, bin1 = num_bins_x,
r2 = r2)
all_centroids_df <- hb_obj$centroids
counts_df <- hb_obj$std_cts
df_bin_centroids <- extract_hexbin_centroids(centroids_df = all_centroids_df,
counts_df = counts_df) |>
dplyr::filter(drop_empty == FALSE)
tr1_object <- tri_bin_centroids(hex_df = df_bin_centroids, x = "c_x", y = "c_y")
tr_from_to_df <- gen_edges(tri_object = tr1_object)
cal_2d_dist(tr_coord_df = tr_from_to_df, start_x = "x_from", start_y = "y_from",
end_x = "x_to", end_y = "y_to", select_vars = c("from", "to", "distance"))

}
