% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2D.R
\name{cal_2D_dist}
\alias{cal_2D_dist}
\title{Calculate 2D Distances Between Points}
\usage{
cal_2D_dist(
  .data,
  start_x = "x_from",
  start_y = "y_from",
  end_x = "x_to",
  end_y = "y_to",
  select_col_vec = c("from", "to", "distance")
)
}
\arguments{
\item{.data}{A data frame containing columns for the x and y coordinates of start and end points.}

\item{start_x}{Column name for the x-coordinate of the starting point.}

\item{start_y}{Column name for the y-coordinate of the starting point.}

\item{end_x}{Column name for the x-coordinate of the ending point.}

\item{end_y}{Column name for the y-coordinate of the ending point.}

\item{select_col_vec}{A character vector specifying the columns to be selected in the resulting data frame.}
}
\value{
A data frame with columns for the starting point, ending point, and calculated distances.
}
\description{
This function calculates the 2D distances between pairs of points in a data frame.
}
\examples{
nldr_df <- s_curve_noise_umap
num_bins <- 7
shape_val <- 1.833091
hexbin_data_object <- extract_hexbin_mean(nldr_df, num_bins, shape_val)
df_bin_centroids <- hexbin_data_object$hexdf_data
tr1_object <- triangulate_bin_centroids(df_bin_centroids, x, y)
tr_from_to_df <- generate_edge_info(triangular_object = tr1_object)
cal_2D_dist(tr_from_to_df, start_x = "x_from", start_y = "y_from",
end_x = "x_to", end_y = "y_to", select_col_vec = c("from", "to", "distance"))

}
