% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{calc_bins_y}
\alias{calc_bins_y}
\title{Calculate the effective number of bins along x-axis and y-axis}
\usage{
calc_bins_y(bin1 = 2, r2)
}
\arguments{
\item{bin1}{Number of bins along the x axis.}

\item{r2}{The ratio of the ranges of the original embedding components.}
}
\value{
A list of numeric values that represents the effective number of
bins along the y axis and width of the hexagon.
}
\description{
This function calculates the effective number of bins along the x and y axes
of a hexagonal grid.
}
\examples{
r2 <- diff(range(s_curve_noise_umap$UMAP2))/diff(range(s_curve_noise_umap$UMAP1))
calc_bins_y(bin1 = 2, r2 = r2)

}
