% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{calculate_effective_x_bins}
\alias{calculate_effective_x_bins}
\title{Calculate the effective number of bins along x-axis}
\usage{
calculate_effective_x_bins(.data, x = UMAP1, cell_area = 1)
}
\arguments{
\item{.data}{A data frame, data frame extension (e.g., a tibble), or a lazy data frame (e.g., from dbplyr or dtplyr).}

\item{x}{A literal variable name representing the x-axis.}

\item{cell_area}{A numeric value that initialise the area of the hexagon.}
}
\value{
A numeric value representing the effective number of bins along x-axis.
}
\description{
This function calculates the effective number of bins along the x-axis of a hexbin plot.
}
\examples{
nldr_df <- s_curve_noise_umap
calculate_effective_x_bins(nldr_df, x = UMAP1, cell_area = 1)
}
