% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2D.R
\name{colour_long_edges}
\alias{colour_long_edges}
\title{Color Long Edges}
\usage{
colour_long_edges(.data, benchmark_value, triangular_object, distance_col)
}
\arguments{
\item{.data}{The data frame containing the edge information.}

\item{benchmark_value}{The threshold value to determine long edges.}

\item{triangular_object}{The triangular object containing the mesh information.}

\item{distance_col}{The column name in `.data` representing the distances.}
}
\value{
A ggplot object with the triangular mesh plot where long edges are colored differently.
}
\description{
This function colors the long edges in a triangular mesh plot based on a benchmark value.
}
\examples{
num_bins_x <- 4
shape_value <- 1.833091
hexbin_data_object <- extract_hexbin_centroids(nldr_df = s_curve_noise_umap,
num_bins = num_bins_x, shape_val = shape_value)
df_bin_centroids <- hexbin_data_object$hexdf_data
tr1_object <- triangulate_bin_centroids(df_bin_centroids, x, y)
tr_from_to_df <- generate_edge_info(triangular_object = tr1_object)
distance_df <- cal_2d_dist(tr_from_to_df)
colour_long_edges(.data = distance_df, benchmark_value = 5.4,
triangular_object = tr1_object, distance_col = "distance")

}
