% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_data_link_plots.R
\name{comb_all_data_model_error}
\alias{comb_all_data_model_error}
\title{Create a dataframe with averaged high-dimensional data and
high-dimensional data, non-linear dimension reduction data, model error data}
\usage{
comb_all_data_model_error(
  highd_data,
  nldr_data,
  model_highd,
  model_2d,
  error_df
)
}
\arguments{
\item{highd_data}{A tibble that contains the high-dimensional data.}

\item{nldr_data}{A tibble that contains the non-linear dimension reduction data.}

\item{model_highd}{A tibble that contains the high-dimensional coordinates of bin centroids.}

\item{model_2d}{The dataset with hexagonal bin centroids.}

\item{error_df}{A tibble that high-dimesional model error.}
}
\value{
A tibble with the average values of the high-dimensional data within
each hexagonal bin and high-dimensional data, non-linear dimension reduction data, model error.
}
\description{
This function combine the average values of high-dimensional data within each
hexagonal bin and high-dimensional data, non-linear dimension reduction data,
model error data.
}
\examples{
df_bin_centroids <- s_curve_obj$s_curve_umap_model_obj$df_bin_centroids
df_bin <- s_curve_obj$s_curve_umap_model_obj$df_bin
model_error <- augment(model_2d = df_bin_centroids, model_highd = df_bin,
highd_data = s_curve_noise_training)
comb_all_data_model_error(highd_data = s_curve_noise_training, nldr_data = s_curve_obj$s_curve_umap_scaled_obj$scaled_nldr,
model_highd = df_bin, model_2d = df_bin_centroids, error_df = model_error)

}
