% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{compute_mean_density_hex}
\alias{compute_mean_density_hex}
\title{Compute Mean Density of Hexagonal Bins}
\usage{
compute_mean_density_hex(
  df_bin_centroids,
  num_bins_x,
  col_std_counts = std_counts,
  col_hb_id = "hexID"
)
}
\arguments{
\item{df_bin_centroids}{A data frame containing information about hexagonal bin centroids,
including the hexagon ID and the standard normalized counts (\code{std_counts}).}

\item{num_bins_x}{The number of bins along the x-axis for the hexagonal grid.}

\item{col_std_counts}{The name of the column containing the standardized counts.}

\item{col_hb_id}{The name of the column containing the hexagonal bin IDs.}
}
\value{
A data frame with an additional column, \code{mean_density}, representing the mean
density of each hexagonal bin based on its neighboring bins.
}
\description{
This function calculates the mean density of hexagonal bins based on their neighboring bins.
}
\examples{
num_bins_x <- 4
shape_value <- 1.833091
hexbin_data_object <- extract_hexbin_centroids(nldr_df = s_curve_noise_umap,
num_bins = num_bins_x, shape_val = shape_value)
df_bin_centroids <- hexbin_data_object$hexdf_data
compute_mean_density_hex(df_bin_centroids, num_bins_x)

}
