% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{compute_mean_density_hex}
\alias{compute_mean_density_hex}
\title{Compute mean density of hexagonal bins}
\usage{
compute_mean_density_hex(df_bin_centroids, bin1)
}
\arguments{
\item{df_bin_centroids}{A tibble that contains information about hexagonal bin centroids,
including the hexagon ID and the standard normalized counts (\code{std_counts}).}

\item{bin1}{The number of bins along the x-axis for the hexagonal grid.}
}
\value{
A tibble contains hexagonal IDs and the mean
density of each hexagonal bin based on its neighboring bins.
}
\description{
This function calculates the mean density of hexagonal bins based on their neighboring bins.
}
\examples{
num_bins_x <- 4
df_bin_centroids <- s_curve_obj$s_curve_umap_model_obj$df_bin_centroids
compute_mean_density_hex(df_bin_centroids, bin1 = num_bins_x)

}
