% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{compute_mean_density_hex}
\alias{compute_mean_density_hex}
\title{Compute mean density of hexagonal bins}
\usage{
compute_mean_density_hex(model_2d, b1 = 30)
}
\arguments{
\item{model_2d}{A tibble that contains information about hexagonal bin centroids,
including the hexagon ID and the standardised counts (\code{w_h}).}

\item{b1}{The number of bins along the x-axis for the hexagonal grid.}
}
\value{
A tibble contains hexagonal IDs and the mean
density of each hexagonal bin based on its neighboring bins.
}
\description{
This function calculates the mean density of hexagonal bins based on their neighboring bins.
}
\examples{
compute_mean_density_hex(model_2d = scurve_model_obj$model_2d, b1 = 30)

}
