% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2d.R
\name{compute_std_counts}
\alias{compute_std_counts}
\title{Compute standardize counts in hexagons}
\usage{
compute_std_counts(data_hb)
}
\arguments{
\item{data_hb}{A tibble with embedding and hexagonal bin IDs.}
}
\value{
A tibble that contains hexagon IDs and the corresponding standardize counts.
}
\description{
This function computes the standardize number of points within each hexagon.
}
\examples{
umap_with_hb_id <- s_curve_obj$s_curve_umap_hb_obj$data_hb_id
compute_std_counts(data_hb = umap_with_hb_id)

}
