% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2d.R
\name{compute_std_counts}
\alias{compute_std_counts}
\title{Compute standardize counts in hexagons}
\usage{
compute_std_counts(data_hb)
}
\arguments{
\item{data_hb}{A tibble with embedding and hexagonal bin IDs.}
}
\value{
A tibble that contains hexagon IDs and the corresponding standardize counts.
}
\description{
This function computes the standardize number of points within each hexagon.
}
\examples{
range_umap2 <- diff(range(s_curve_noise_umap$UMAP2))
all_centroids_df <- gen_centroids(bin1 = 3, s1 = -0.1, s2 = -0.1,
r2 = range_umap2)
umap_with_hb_id <- assign_data(data = s_curve_noise_umap_scaled,
centroid_df = all_centroids_df)
compute_std_counts(data_hb = umap_with_hb_id)

}
