% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highD.R
\name{compute_weights}
\alias{compute_weights}
\title{Compute Weights for Hexagonal Binning}
\usage{
compute_weights(nldr_df, hb_object)
}
\arguments{
\item{nldr_df}{A data frame containing 2D embeddings without a unique identifier column (ID).}

\item{hb_object}{A hexbin object containing the hexagonal binning information.}
}
\value{
A data frame with weights calculated for each hexagonal bin.
}
\description{
This function computes weights for hexagonal binning based on the average values of each bin and the distances from these averages.
}
\examples{
num_bins_x <- 4
shape_value <- 1.833091
hexbin_data_object <- extract_hexbin_centroids(nldr_df = s_curve_noise_umap,
num_bins = num_bins_x, shape_val = shape_value)
hexdf_data <- hexbin_data_object$hexdf_data
hb_obj <- hexbin_data_object$hb_data
compute_weights(nldr_df = s_curve_noise_umap |> dplyr::select(-ID), hb_object = hb_obj)

}
