% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highD.R
\name{compute_weights}
\alias{compute_weights}
\title{Compute Weights for Hexagonal Binning}
\usage{
compute_weights(nldr_df, hb_object)
}
\arguments{
\item{nldr_df}{A data frame containing UMAP coordinates and a unique identifier column (ID).}

\item{hb_object}{A hexbin object containing the hexagonal binning information.}
}
\value{
A data frame with weights calculated for each hexagonal bin.
}
\description{
This function computes weights for hexagonal binning based on the average values of each bin and the distances from these averages.
}
\examples{
nldr_df <- s_curve_noise_umap
num_bins <- 8
shape_val <- 2.031141
result <- extract_hexbin_centroids(nldr_df, num_bins, shape_val)
hexdf_data <- result$hexdf_data
hb_obj <- result$hb_data
compute_weights(nldr_df, hb_obj)

}
