% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2D.R
\name{create_hexbin}
\alias{create_hexbin}
\title{Generate the Hexagonal}
\usage{
create_hexbin(
  .data,
  nldr_df,
  embedding_1,
  embedding_2,
  num_bins = 30,
  shape_val = 1,
  apply_pca = TRUE
)
}
\arguments{
\item{apply_pca}{}
}
\value{
an S4 object of class "hexbin".

hb: an S4 object of class "hexbin".
}
\description{
Generate the Hexagonal
}
