% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_2d.R
\name{extract_hexbin_centroids}
\alias{extract_hexbin_centroids}
\title{Extract hexagonal bin centroids coordinates and the corresponding standardise counts.}
\usage{
extract_hexbin_centroids(centroids_df, counts_df)
}
\arguments{
\item{centroids_df}{A tibble that contains all hexagonal bin centroid
coordinates with hexagon IDs.}

\item{counts_df}{A tibble that contains hexagon IDs with the standardise
number of points within each hexagon.}
}
\value{
A tibble contains hexagon ID, centroid coordinates, and standardise counts.
}
\description{
Extract hexagonal bin centroids coordinates and the corresponding standardise counts.
}
\examples{
all_centroids_df <- s_curve_obj$s_curve_umap_hb_obj$centroids
counts_df <- s_curve_obj$s_curve_umap_hb_obj$std_cts
extract_hexbin_centroids(centroids_df = all_centroids_df,
counts_df = counts_df)

}
