% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2D.R
\name{extract_hexbin_mean}
\alias{extract_hexbin_mean}
\title{Extract hexagonal bin mean and related information from non-linear dimensionality reduction data.}
\usage{
extract_hexbin_mean(nldr_df, num_bins, shape_val = 1, x = UMAP1, y = UMAP2)
}
\arguments{
\item{nldr_df}{Non-linear dimensionality reduction data frame containing 2D coordinates.}

\item{num_bins}{Number of bins along the x-axis for hexagon binning.}

\item{shape_val}{The value of the shape parameter for hexagon binning.}
}
\value{
A list containing the hexagonal bin mean data frame and the hexbin object.
}
\description{
Extract hexagonal bin mean and related information from non-linear dimensionality reduction data.
}
\examples{
# Example usage of extract_hexbin_centroids function
nldr_df <- tibble::tibble(ID = 1:100, UMAP1 = rnorm(100), UMAP2 = rnorm(100))
num_bins <- 20
shape_val <- 0.8
result <- extract_hexbin_mean(nldr_df, num_bins, shape_val)
hexdf_data <- result$hexdf_data
hb_data <- result$hb_data
print(hexdf_data)
print(hb_data)


}
