% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2d.R
\name{extract_hexbin_mean}
\alias{extract_hexbin_mean}
\title{Extract hexagonal bin mean coordinates and the corresponding standardize counts.}
\usage{
extract_hexbin_mean(data_hb, counts_df)
}
\arguments{
\item{data_hb}{A tibble with embedding components and hexagonal bin IDs.}

\item{counts_df}{A tibble that contains hexagon IDs with the standardise
number of points within each hexagon.}
}
\value{
A tibble contains hexagon ID, bin mean coordinates, and standardize counts.
}
\description{
Extract hexagonal bin mean coordinates and the corresponding standardize counts.
}
\examples{
r2 <- diff(range(s_curve_noise_umap$UMAP2))/diff(range(s_curve_noise_umap$UMAP1))
num_bins_x <- 3
hb_obj <- hex_binning(data = s_curve_noise_umap_scaled, bin1 = num_bins_x,
r2 = r2)
umap_with_hb_id <- hb_obj$data_hb_id
counts_df <- hb_obj$std_cts
extract_hexbin_mean(data_hb = umap_with_hb_id, counts_df = counts_df)

}
