% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2d.R
\name{extract_hexbin_mean}
\alias{extract_hexbin_mean}
\title{Extract hexagonal bin mean coordinates and the corresponding standardize counts.}
\usage{
extract_hexbin_mean(nldr_df_with_hex_id, counts_df)
}
\arguments{
\item{nldr_df_with_hex_id}{A data frame with 2D embeddings and hexagonal bin IDs.}

\item{counts_df}{A data frame contains hexagon IDs with the standardize number of points within each hexagon.}
}
\value{
A tibble contains hexagon ID, bin mean coordinates, and standardize counts.
}
\description{
Extract hexagonal bin mean coordinates and the corresponding standardize counts.
}
\examples{
num_bins_list <- calc_bins(data = s_curve_noise_umap_scaled, x = "UMAP1", y = "UMAP2",
hex_size = 0.2, buffer_x = 0.346, buffer_y = 0.3)
num_bins_x <- num_bins_list$num_x
num_bins_y <- num_bins_list$num_y
hb_obj <- hex_binning(data = s_curve_noise_umap_scaled,
x = "UMAP1", y = "UMAP2", num_bins_x = num_bins_x,
num_bins_y = num_bins_y, x_start = NA, y_start = NA, buffer_x = 0.346,
buffer_y = 0.3, hex_size = 0.2, col_start = "UMAP")
all_centroids_df <- hb_obj$centroids
counts_df <- hb_obj$std_cts
umap_with_hb_id <- assign_data(data = s_curve_noise_umap_scaled,
centroid_df = all_centroids_df, col_start = "UMAP")
extract_hexbin_mean(nldr_df_with_hex_id = umap_with_hb_id, counts_df = counts_df)

}
