% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highD.R
\name{find_benchmark_value}
\alias{find_benchmark_value}
\title{Find Benchmark Value}
\usage{
find_benchmark_value(.data, distance_col)
}
\arguments{
\item{.data}{The data frame containing the distances.}

\item{distance_col}{The name of the column containing the distances.}
}
\value{
The benchmark value, which is the first largest difference in the distance column.
}
\description{
This function finds the benchmark value to remove long edges based on the differences in a distance column.
}
\examples{
df <- tibble::tribble(
  ~x, ~y, ~distance,
  1, 2, 0.5,
  4, 5, 0.7,
  7, 8, 0.2
)
find_benchmark_value(df, distance_col = "distance")

}
