% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{find_benchmark_value}
\alias{find_benchmark_value}
\title{Find Benchmark Value To Remove Long Edges}
\usage{
find_benchmark_value(.data, distance_col)
}
\arguments{
\item{.data}{The data frame containing the distances.}

\item{distance_col}{The name of the column containing the distances.}
}
\value{
The benchmark value, which is the first largest difference in the distance column.
}
\description{
This function finds the benchmark value to remove long edges based on the differences in a distance column.
}
\examples{
nldr_df <- s_curve_noise_umap
num_bins <- 8
shape_val <- 2.031141
hexbin_data_object <- extract_hexbin_mean(nldr_df, num_bins, shape_val)
df_bin_centroids <- hexbin_data_object$hexdf_data
tr1_object <- triangulate_bin_centroids(df_bin_centroids, x, y)
tr_from_to_df <- generate_edge_info(triangular_object = tr1_object)
distance_df <- cal_2D_dist(tr_from_to_df)
find_benchmark_value(distance_df, distance_col = "distance")

}
