% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/04_parameters.R
\name{find_lg_benchmark}
\alias{find_lg_benchmark}
\title{Compute a benchmark value to remove long edges}
\usage{
find_lg_benchmark(distance_edges, distance_col)
}
\arguments{
\item{distance_edges}{The tibble contains the distances.}

\item{distance_col}{The name of the column containing the distances.}
}
\value{
The benchmark value, which is the first largest difference in the distance column.
}
\description{
This function finds the benchmark value to remove long edges based on
the differences in a distance column.
}
\examples{
distance_df <- s_curve_obj$s_curve_umap_model_distance_df
find_lg_benchmark(distance_edges = distance_df, distance_col = "distance")

}
