% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{find_low_dens_hex}
\alias{find_low_dens_hex}
\title{Find low-density Hexagons}
\usage{
find_low_dens_hex(model_2d, b1 = 5, benchmark_mean_dens = 0.05)
}
\arguments{
\item{model_2d}{The tibble that contains all hexagonal bin centroids.}

\item{b1}{Number of bins along the x-axis for hexagon binning.}

\item{benchmark_mean_dens}{A numeric value that contains}
}
\value{
A vector containing the IDs of hexagons to be removed after investigating their neighboring bins.
}
\description{
This function identifies hexagons with low density based on the mean density
of their neighboring hexagons.
}
\examples{
find_low_dens_hex(model_2d = scurve_model_obj$model_2d, b1 = 5,
benchmark_mean_dens = 0.05)

}
