% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2d.R
\name{find_non_empty_bins}
\alias{find_non_empty_bins}
\title{Find the number of bins required to achieve required number of non-empty bins.}
\usage{
find_non_empty_bins(data, non_empty_bins, s1 = -0.1, s2 = -0.1, r2)
}
\arguments{
\item{data}{A tibble that contains the embedding.}

\item{non_empty_bins}{The desired number of non-empty bins.}

\item{s1}{The x-coordinate of the hexagonal grid starting point.}

\item{s2}{The y-coordinate of the hexagonal grid starting point.}

\item{r2}{The range of the original second embedding component.}
}
\value{
The number of bins along the x and y axes
needed to achieve a specific number of non-empty bins.
}
\description{
This function determines the number of bins along the x and y axes
to obtain a specific number of non-empty bins.
}
\examples{
r2 <- diff(range(s_curve_noise_umap$UMAP2))/diff(range(s_curve_noise_umap$UMAP1))
find_non_empty_bins(data = s_curve_noise_umap_scaled, non_empty_bins = 6,
r2 = r2)

}
