% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2d.R
\name{find_pts}
\alias{find_pts}
\title{Find points in hexagonal bins}
\usage{
find_pts(data_hb)
}
\arguments{
\item{data_hb}{A data frame with data, ID and hexagonal bin IDs.}
}
\value{
A tibble with hexagonal bin IDs and the corresponding points.
}
\description{
This function maps points to their corresponding hexagonal bins.
}
\examples{
range_umap2 <- diff(range(s_curve_noise_umap$UMAP2))
all_centroids_df <- gen_centroids(bin1 = 3, s1 = -0.1, s2 = -0.1,
r2 = range_umap2)
umap_with_hb_id <- assign_data(data = s_curve_noise_umap_scaled,
centroid_df = all_centroids_df)
find_pts(data_hb = umap_with_hb_id)

}
