% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_2d.R
\name{find_pts}
\alias{find_pts}
\title{Find points in hexagonal bins}
\usage{
find_pts(data_hb)
}
\arguments{
\item{data_hb}{A data frame with data, ID and hexagonal bin IDs.}
}
\value{
A tibble with hexagonal bin IDs and the corresponding points.
}
\description{
This function maps points to their corresponding hexagonal bins.
}
\examples{
umap_with_hb_id <- s_curve_obj$s_curve_umap_hb_obj$data_hb_id
find_pts(data_hb = umap_with_hb_id)

}
