% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2d.R
\name{find_pts}
\alias{find_pts}
\title{Find points in hexagonal bins}
\usage{
find_pts(scaled_nldr_hexid)
}
\arguments{
\item{scaled_nldr_hexid}{A tibble that contains the scaled embedding with hexagonal bin IDs.}
}
\value{
A tibble with hexagonal bin IDs (\code{h}) and the corresponding points.
}
\description{
This function maps points to their corresponding hexagonal bins.
}
\examples{
umap_with_hb_id <- scurve_model_obj$hb_obj$data_hb_id
find_pts(scaled_nldr_hexid = umap_with_hb_id)

}
