% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2D.R
\name{find_pts_in_hexbins}
\alias{find_pts_in_hexbins}
\title{Find Points in Hexagonal Bins}
\usage{
find_pts_in_hexbins(full_grid_with_hexbin_id, UMAP_data_with_hb_id)
}
\arguments{
\item{full_grid_with_hexbin_id}{A data frame with hexagonal bin IDs and coordinates.}

\item{UMAP_data_with_hb_id}{A data frame with UMAP data and hexagonal bin IDs.}
}
\value{
A data frame with hexagonal bin IDs and the corresponding points.
}
\description{
This function maps points to their corresponding hexagonal bins based on the provided data frames.
}
\examples{
nldr_df <- s_curve_noise_umap
num_bins <- 8
shape_val <- 2.031141
hexbin_data_object <- extract_hexbin_mean(nldr_df, num_bins, shape_val)
df_bin_centroids <- hexbin_data_object$hexdf_data
full_centroid_df <- generate_full_grid_centroids(df_bin_centroids)
hex_grid <- full_hex_grid(full_centroid_df)
full_grid_with_hexbin_id <- map_hexbin_id(full_centroid_df, df_bin_centroids)
UMAP_data_with_hb_id <- nldr_df |> dplyr::mutate(hb_id = hexbin_data_object$hb_data@cID)
find_pts_in_hexbins(full_grid_with_hexbin_id, UMAP_data_with_hb_id)

}
