% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{fit_highd_model}
\alias{fit_highd_model}
\title{Construct the 2D model and lift into high-D}
\usage{
fit_highd_model(
  highd_data,
  nldr_data,
  bin1 = 4,
  r2,
  q = 0.1,
  is_bin_centroid = TRUE
)
}
\arguments{
\item{highd_data}{A tibble that contains the training high-dimensional data.}

\item{nldr_data}{A tibble that contains embedding with a unique identifier.}

\item{bin1}{Number of bins along the x axis.}

\item{r2}{The ratio of the ranges of the original embedding components.}

\item{q}{The buffer amount as proportion of data range.}
}
\value{
A list containing the data frame with high-dimensional coordinates
for 2D bin centroids (\code{df_bin}) and the data frame containing
information about hexagonal bin centroids (\code{df_bin_centroids}) in 2D.
}
\description{
This function fits a high-dimensional model using hexagonal bins and provides options
to customize the modeling process, including the choice of bin centroids or bin means,
removal of low-density hexagons, and averaging of high-dimensional data.
}
\examples{
scurve_umap_scaled_obj <- s_curve_obj$s_curve_umap_scaled_obj
lim1 <- scurve_umap_scaled_obj$lim1
lim2 <- scurve_umap_scaled_obj$lim2
r2 <- diff(lim2)/diff(lim1)
fit_highd_model(highd_data = s_curve_noise_training,
nldr_data = s_curve_noise_umap_scaled, bin1 = 4, r2 = r2)

}
