% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{fit_highd_model}
\alias{fit_highd_model}
\title{Construct the 2D model and lift into high-D}
\usage{
fit_highd_model(
  training_data,
  emb_df,
  bin1 = 4,
  r2,
  q = 0.1,
  is_bin_centroid = TRUE,
  is_rm_lwd_hex = FALSE,
  benchmark_to_rm_lwd_hex = NULL,
  col_start_highd = "x"
)
}
\arguments{
\item{training_data}{A tibble that contains the training high-dimensional data.}

\item{emb_df}{A tibble that contains embedding with a unique identifier.}

\item{bin1}{Number of bins along the x axis.}

\item{r2}{The ratio of the ranges of the original embedding components.}

\item{q}{The buffer amount as proportion of data range.}

\item{is_bin_centroid}{Logical, indicating whether to use bin centroids (default is TRUE).}

\item{is_rm_lwd_hex}{Logical, indicating whether to remove low-density hexagons
(default is FALSE).}

\item{benchmark_to_rm_lwd_hex}{The benchmark value to remove low-density hexagons.}

\item{col_start_highd}{The text prefix for columns in the high-dimensional data.}
}
\value{
A list containing the data frame with high-dimensional coordinates
for 2D bin centroids (\code{df_bin}) and the data frame containing
information about hexagonal bin centroids (\code{df_bin_centroids}) in 2D.
}
\description{
This function fits a high-dimensional model using hexagonal bins and provides options
to customize the modeling process, including the choice of bin centroids or bin means,
removal of low-density hexagons, and averaging of high-dimensional data.
}
\examples{
r2 <- diff(range(s_curve_noise_umap$UMAP2))/diff(range(s_curve_noise_umap$UMAP1))
fit_highd_model(training_data = s_curve_noise_training,
emb_df = s_curve_noise_umap_scaled, bin1 = 4, r2 = r2,
col_start_highd = "x")

}
