% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2D.R
\name{full_hex_grid}
\alias{full_hex_grid}
\title{Generate Hexagonal Coordinates}
\usage{
full_hex_grid(hexdf_data)
}
\arguments{
\item{hexdf_data}{The dataset with all hexbin ID and centroid coordinates.}
}
\value{
A tibble containing the coordinates of hexagons based on hexagonal centroids.
}
\description{
This function generates the coordinates of hexagons after passing hexagonal centroids.
}
\examples{
num_bins_x <- 4
shape_value <- 1.833091
hexbin_data_object <- extract_hexbin_centroids(nldr_df = s_curve_noise_umap,
num_bins = num_bins_x, shape_val = shape_value)
df_bin_centroids <- hexbin_data_object$hexdf_data
full_centroid_df <- generate_full_grid_centroids(df_bin_centroids)
full_hex_grid(full_centroid_df)

}
