% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2D.R
\name{full_hex_grid}
\alias{full_hex_grid}
\title{Generate Hexagonal Coordinates}
\usage{
full_hex_grid(hexdf_data)
}
\arguments{
\item{hexdf_data}{The dataset with all hexbin ID and centroid coordinates.}
}
\value{
A tibble containing the coordinates of hexagons based on hexagonal centroids.
}
\description{
This function generates the coordinates of hexagons after passing hexagonal centroids.
}
\examples{
nldr_df <- s_curve_noise_umap
num_bins <- 7
shape_val <- 1.833091
set.seed(123)
hexbin_data_object <- extract_hexbin_mean(nldr_df, num_bins, shape_val)
df_bin_centroids <- hexbin_data_object$hexdf_data
df_bin_centroids_sample <- df_bin_centroids |> dplyr::slice_sample(n = 5)
full_centroid_df <- generate_full_grid_centroids(df_bin_centroids)
full_hex_grid(full_centroid_df)

}
