% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2d.R
\name{gen_centroids}
\alias{gen_centroids}
\title{Generate centroid coordinate}
\usage{
gen_centroids(nldr_obj, b1 = 4, q = 0.1)
}
\arguments{
\item{b1}{Number of bins along the x axis.}

\item{q}{The buffer amount as proportion of data range.}

\item{nldr_data}{A tibble that contains embedding components in the first and second columns.}
}
\value{
A tibble contains hexIDs, x and y coordinates (h, c_x, c_y respectively)
of all hexagon bin centroids.
}
\description{
This function generates all possible centroids in the hexagonal grid.
}
\examples{
gen_centroids(nldr_obj = scurve_umap_obj, b1 = 4, q = 0.1)

}
