% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{gen_diffbin1_errors}
\alias{gen_diffbin1_errors}
\title{Generate erros and MSE for different bin widths}
\usage{
gen_diffbin1_errors(highd_data, nldr_data, hd_thresh = 1, bin1_vec = NULL)
}
\arguments{
\item{highd_data}{A tibble that contains the high-dimensional data with a unique identifier.}

\item{nldr_data}{A tibble that contains the embedding with a unique identifier.}

\item{hd_thresh}{(default: 1) A numeric value using to filter high-density hexagons.}

\item{bin1_vec}{A numeric vector contains the range of b1 values.}
}
\value{
A tibble containing the augmented data with predictions,
error metrics, and absolute error metrics.
}
\description{
This function augments a dataset with predictions and error metrics obtained
from a nonlinear dimension reduction (NLDR) model.
}
\examples{
scurve_sample <- scurve |> head(100)
scurve_umap_sample <- scurve_umap |> head(100)
gen_diffbin1_errors(highd_data = scurve_sample, nldr_data = scurve_umap_sample)

}
