% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_2d.R
\name{gen_edges}
\alias{gen_edges}
\title{Generate edge information}
\usage{
gen_edges(tri_object)
}
\arguments{
\item{tri_object}{The triangular object from which to generate edge information.}
}
\value{
A tibble that contains the edge information, including the from-to
relationships and the corresponding x and y coordinates.
}
\description{
This function generates edge information from a given triangular object,
including the coordinates of the vertices and the from-to relationships
between the vertices.
}
\examples{
tr1_object <- s_curve_obj$s_curve_umap_model_tr1_object
gen_edges(tri_object = tr1_object)

}
