% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_2d.R
\name{gen_hex_coord}
\alias{gen_hex_coord}
\title{Generate hexagonal polygon coordinates}
\usage{
gen_hex_coord(centroids_df, a1)
}
\arguments{
\item{centroids_df}{The dataset with all hexagonal bin IDs
and centroid coordinates.}

\item{a1}{The width of the hexagon.}
}
\value{
A tibble contains polygon id, x and y coordinates (hex_poly_id, x,
and y respectively) of hexagons.
}
\description{
This function generates the coordinates of hexagons after passing
hexagonal centroids.
}
\examples{
width <- s_curve_obj$s_curve_umap_hb_obj$a1
all_centroids_df <- s_curve_obj$s_curve_umap_hb_obj$centroids
gen_hex_coord(centroids_df = all_centroids_df, a1 = width)

}
