% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2d.R
\name{gen_hex_coord}
\alias{gen_hex_coord}
\title{Generate hexagonal polygon coordinates}
\usage{
gen_hex_coord(centroids_df, a1)
}
\arguments{
\item{centroids_df}{The dataset with all hexagonal bin IDs
and centroid coordinates.}

\item{a1}{The width of the hexagon.}
}
\value{
A tibble contains polygon id, x and y coordinates (hex_poly_id, x,
and y respectively) of hexagons.
}
\description{
This function generates the coordinates of hexagons after passing
hexagonal centroids.
}
\examples{
r2 <- diff(range(s_curve_noise_umap$UMAP2))/diff(range(s_curve_noise_umap$UMAP1))
num_bins_list <- calc_bins_y(bin1 = 4, r2 = r2, q = 0.1)
width <- num_bins_list$a1
all_centroids_df <- gen_centroids(bin1 = 4, r2 = r2, q = 0.1)
gen_hex_coord(centroids_df = all_centroids_df, a1 = width)

}
