% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2d.R
\name{gen_hex_coord}
\alias{gen_hex_coord}
\title{Generate hexagonal polygon coordinates}
\usage{
gen_hex_coord(centroids_data, a1)
}
\arguments{
\item{centroids_data}{The dataset with all hexagonal bin IDs
and centroid coordinates.}

\item{a1}{The width of the hexagon.}
}
\value{
A tibble contains hexagon id (\code{h}),
x and y coordinates (\code{x}, \code{y}) of hexagons.
}
\description{
This function generates the coordinates of hexagons after passing
hexagonal centroids.
}
\examples{
width <- scurve_model_obj$hb_obj$a1
all_centroids_df <- scurve_model_obj$hb_obj$centroids
gen_hex_coord(centroids_data = all_centroids_df, a1 = width)

}
