% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preprocessing.R
\name{gen_scaled_data}
\alias{gen_scaled_data}
\title{Scaling the NLDR data}
\usage{
gen_scaled_data(nldr_data)
}
\arguments{
\item{nldr_data}{A tibble that contains embedding components in the first and second columns.}
}
\value{
A list of a tibble contains scaled first and second columns NLDR data,
and numeric vectors representing the limits of the original NLDR data.
}
\description{
This function scales first and second columns.
}
\examples{
gen_scaled_data(nldr_data = scurve_umap)

}
