% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2D.R
\name{generate_edge_info}
\alias{generate_edge_info}
\title{Generate Edge Information}
\usage{
generate_edge_info(triangular_object)
}
\arguments{
\item{triangular_object}{The triangular object from which to generate edge information.}
}
\value{
A data frame containing the edge information, including the from-to relationships and the corresponding x and y coordinates.
}
\description{
This function generates edge information from a given triangular object, including the coordinates of the vertices and the from-to relationships between the vertices.
}
\examples{
tr_obj <- tripack::tri.mesh(x = c(1, 2, 3), y = c(4, 5, 6))
generate_edge_info(tr_obj)

}
