% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2D.R
\name{generate_edge_info}
\alias{generate_edge_info}
\title{Generate Edge Information}
\usage{
generate_edge_info(triangular_object)
}
\arguments{
\item{triangular_object}{The triangular object from which to generate edge information.}
}
\value{
A data frame containing the edge information, including the from-to
relationships and the corresponding x and y coordinates.
}
\description{
This function generates edge information from a given triangular object,
including the coordinates of the vertices and the from-to relationships between the vertices.
}
\examples{
num_bins_x <- 4
shape_value <- 1.833091
hexbin_data_object <- extract_hexbin_centroids(nldr_df = s_curve_noise_umap,
num_bins = num_bins_x, shape_val = shape_value)
df_bin_centroids <- hexbin_data_object$hexdf_data
tr1_object <- triangulate_bin_centroids(df_bin_centroids, "x", "y")
generate_edge_info(triangular_object = tr1_object)

}
