% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{generate_eval_df}
\alias{generate_eval_df}
\title{Generate Evaluation Data Frame}
\usage{
generate_eval_df(
  data,
  prediction_df,
  df_bin_centroids,
  df_bin,
  num_bins,
  col_start = "x"
)
}
\arguments{
\item{data}{The data set containing high-dimensional data along with an unique identifier.}

\item{prediction_df}{The data set with 2D embeddings, IDs, and predicted hexagonal IDs.}

\item{df_bin_centroids}{The data set with coordinates of hexagonal bin centroids.}

\item{df_bin}{The data set with averaged/weighted high-dimensional data.}

\item{num_bins}{Number of bins along the x-axis for hexagon binning.}

\item{col_start}{The text that begin the column name of the high-D data}
}
\value{
A tibble containing evaluation metrics based on the provided inputs.
}
\description{
This function generates an evaluation data frame based on the provided data and predictions.
}
\examples{
num_bins_x <- 4
shape_value <- 1.833091
hexbin_data_object <- extract_hexbin_mean(nldr_df = s_curve_noise_umap, num_bins_x,
shape_val = shape_value)
df_bin_centroids <- hexbin_data_object$hexdf_data
UMAP_data_with_hb_id <- s_curve_noise_umap |> dplyr::mutate(hb_id = hexbin_data_object$hb_data@cID)
df_all <- dplyr::bind_cols(s_curve_noise_training |> dplyr::select(-ID), UMAP_data_with_hb_id)
df_bin <- avg_highD_data(df_all)
pred_df_test <- predict_2d_embeddings(test_data = s_curve_noise_training,
df_bin_centroids = df_bin_centroids,
df_bin = df_bin, type_NLDR = "UMAP")
generate_eval_df(data = s_curve_noise, prediction_df = pred_df_test,
df_bin_centroids = df_bin_centroids, df_bin = df_bin, num_bins = num_bins_x, col_start = "x")

}
