% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{generate_eval_df}
\alias{generate_eval_df}
\title{Generate Evaluation Data Frame}
\usage{
generate_eval_df(
  data,
  prediction_df,
  df_bin_centroids,
  df_bin,
  num_bins,
  col_start = "x"
)
}
\arguments{
\item{data}{The data set containing high-dimensional data along with IDs.}

\item{prediction_df}{The data set with 2D embeddings, IDs, and predicted hexagonal IDs.}

\item{df_bin_centroids}{The data set with coordinates of hexagonal bin centroids.}

\item{df_bin}{The data set with averaged/weighted high-dimensional data.}

\item{num_bins}{Number of bins along the x-axis for hexagon binning.}

\item{col_start}{The text that begin the column name of the high-D data}
}
\value{
A tibble containing evaluation metrics based on the provided inputs.
}
\description{
This function generates an evaluation data frame based on the provided data and predictions.
}
\examples{
data <- s_curve_noise
nldr_df <- s_curve_noise_umap
nldr_df_test <- s_curve_noise_umap_predict
training_data <- s_curve_noise_training
num_bins <- 8
shape_val <- 2.031141
pred_df_test_object <- predict_hex_id(training_data = training_data,
nldr_df = nldr_df, nldr_df_test = nldr_df_test, num_bins = num_bins, shape_val = shape_val)
pred_df_test <- pred_df_test_object$pred_data
centroid_df_test <- pred_df_test_object$df_bin_centroids
avg_df_test <- pred_df_test_object$df_bin
generate_eval_df(data = data, prediction_df = pred_df_test,
df_bin_centroids = centroid_df_test, df_bin = avg_df_test, num_bins = num_bins, col_start = "x")

}
