% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2D.R
\name{generate_full_grid_centroids}
\alias{generate_full_grid_centroids}
\title{Generate Full Grid Centroids}
\usage{
generate_full_grid_centroids(hexdf_data)
}
\arguments{
\item{hexdf_data}{The dataset with hexbin ID and centroid coordinates.}
}
\value{
A tibble containing all possible centroids in the full grid.
}
\description{
This function generates all possible centroids in the full grid based on hexbin data.
}
\examples{
nldr_df <- readRDS(paste0(here::here(), "/quollr/data-raw/s_curve_noise_umap.rds"))
num_bins <- 8
shape_val <- 2.031141
hexbin_data_object <- extract_hexbin_mean(nldr_df, num_bins, shape_val)
df_bin_centroids <- hexbin_data_object$hexdf_data
generate_full_grid_centroids(df_bin_centroids)

}
