% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2D.R
\name{generate_full_grid_centroids}
\alias{generate_full_grid_centroids}
\title{Generate Full Grid Centroids}
\usage{
generate_full_grid_centroids(hexdf_data)
}
\arguments{
\item{hexdf_data}{The dataset with hexbin ID and centroid coordinates.}
}
\value{
A tibble containing all possible centroids in the full grid.
}
\description{
This function generates all possible centroids in the full grid based on hexbin data.
}
\examples{
hexdf_data <- tibble::tibble(
  x = c(0, 1, 0.5),
  y = c(0, 0, 1)
)
generate_full_grid_centroids(hexdf_data)

}
