% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2D.R
\name{generate_full_grid_info}
\alias{generate_full_grid_info}
\title{Generate Full Grid Information Data Frame}
\usage{
generate_full_grid_info(df_bin_centroids)
}
\arguments{
\item{df_bin_centroids}{A data frame with hexagonal bin centroids.}
}
\value{
A data frame with columns "x", "y", "id", "c_x", "c_y", "hexID", "counts", "std_counts", and "polygon_id".
}
\description{
This function generates a data frame containing coordinates and identifiers for hexagons,
along with additional information like counts and polygon IDs.
}
\examples{
df_bin_centroids <- tibble::tibble(
  hexID = c(1, 2, 3),
  x = c(1.5, 2.5, 3.5),
  y = c(4.5, 5.5, 6.5)
)
generate_full_grid_info(df_bin_centroids)

}
