% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{glance}
\alias{glance}
\title{Generate evaluation metrics}
\usage{
glance(
  df_bin_centroids,
  df_bin,
  training_data,
  newdata = NULL,
  type_NLDR,
  col_start = "x"
)
}
\arguments{
\item{df_bin_centroids}{Centroid coordinates of hexagonal bins in 2D space.}

\item{df_bin}{Centroid coordinates of hexagonal bins in high dimensions.}

\item{training_data}{Training data used to fit the model.}

\item{newdata}{Data to be augmented with predictions and error metrics.
If NULL, the training data is used (default is NULL).}

\item{type_NLDR}{The type of non-linear dimensionality reduction (NLDR) used.}

\item{col_start}{The text that begin the column name of the high-dimensional data.}
}
\value{
A tibble contains Error, and MSE values.
}
\description{
This function generates an evaluation data frame based on the provided data and predictions.
}
\examples{
df_bin_centroids <- s_curve_obj$s_curve_umap_model_distance_df$df_bin_centroids
df_bin <- s_curve_obj$s_curve_umap_model_distance_df$df_bin
glance(df_bin_centroids = df_bin_centroids, df_bin = df_bin,
training_data = s_curve_noise_training, newdata = NULL, type_NLDR = "UMAP",
col_start = "x")

}
