% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{glance}
\alias{glance}
\title{S3 generic for glance}
\usage{
glance(x, ...)
}
\arguments{
\item{x}{An object to compute evaluation metrics for.}

\item{...}{Additional arguments passed to methods.}
}
\description{
This is a generic function for computing evaluation metrics on different objects.
}
