% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2d.R
\name{hex_binning}
\alias{hex_binning}
\title{Hexagonal binning}
\usage{
hex_binning(data, bin1 = 2, s1 = -0.1, s2 = -0.1, r2)
}
\arguments{
\item{data}{A tibble that contains embedding components.}

\item{bin1}{Number of bins along the x axis.}

\item{s1}{The x-coordinate of the hexagonal grid starting point.}

\item{s2}{The y-coordinate of the hexagonal grid starting point.}

\item{r2}{The range of the original second embedding component.}
}
\value{
A object that contains numeric vector that contains bins along the x and y axes respectively,
all hexagonal bin centroids (centroids),
hexagonal coordinates of the full grid(hex_poly),
embedding components with their corresponding hexagon IDs (data_hb_id),
hex bins with their corresponding standardise counts (std_cts),
total number of hex bins(tot_bins), number of non-empty hex bins (non_bins)
and points within each hexagon (pts_bins).
}
\description{
This function generates the hexagonal object.
}
\examples{
range_umap2 <- diff(range(s_curve_noise_umap$UMAP2))
num_bins_x <- 3
hex_binning(data = s_curve_noise_umap_scaled, bin1 = num_bins_x, s1 = -0.1,
s2 = -0.1, r2 = range_umap2)

}
