% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_2d.R
\name{hex_binning}
\alias{hex_binning}
\title{Hexagonal binning}
\usage{
hex_binning(data, bin1 = 4, r2, q = 0.1)
}
\arguments{
\item{data}{A tibble that contains embedding components.}

\item{bin1}{Number of bins along the x axis.}

\item{r2}{The ratio of the ranges of the original embedding components.}

\item{q}{The buffer amount as proportion of data range.}
}
\value{
A object that contains numeric vector that contains binwidths (a1),
vertical distance (a2), bins along the x and y axes respectively (bins),
numeric vector that contains hexagonal
starting point coordinates all hexagonal bin centroids (centroids),
hexagonal coordinates of the full grid(hex_poly),
embedding components with their corresponding hexagon IDs (data_hb_id),
hex bins with their corresponding standardise counts (std_cts),
total number of hex bins(tot_bins), number of non-empty hex bins (non_bins)
and points within each hexagon (pts_bins).
}
\description{
This function generates the hexagonal object.
}
\examples{
scurve_umap_scaled_obj <- s_curve_obj$nldr_scaled_obj
lim1 <- scurve_umap_scaled_obj$lim1
lim2 <- scurve_umap_scaled_obj$lim2
r2 <- diff(lim2)/diff(lim1)
num_bins_x <- 4
hex_binning(data = s_curve_noise_umap_scaled, bin1 = num_bins_x,
r2 = r2, q = 0.1)

}
