% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2D.R
\name{map_hexbin_id}
\alias{map_hexbin_id}
\title{Map Hexagon IDs to Centroids in the Full Grid}
\usage{
map_hexbin_id(full_centroid_df, df_bin_centroids)
}
\arguments{
\item{full_centroid_df}{Data frame containing centroid coordinates of the full grid.}

\item{df_bin_centroids}{Data frame containing hexagon IDs and their centroids.}
}
\value{
A data frame with columns 'x', 'y', 'hexID', and 'counts' representing hexagon centroids and counts.
}
\description{
This function generates a data frame with hexagon IDs mapped to their centroids in the full grid.
}
\examples{
nldr_df <- s_curve_noise_umap
num_bins <- 8
shape_val <- 2.031141
hexbin_data_object <- extract_hexbin_mean(nldr_df, num_bins, shape_val)
df_bin_centroids <- hexbin_data_object$hexdf_data
full_centroid_df <- generate_full_grid_centroids(df_bin_centroids)
map_hexbin_id(full_centroid_df, df_bin_centroids)

}
