% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2D.R
\name{map_polygon_id}
\alias{map_polygon_id}
\title{Map Polygon ID to Hexagon Coordinates}
\usage{
map_polygon_id(full_grid_with_hexbin_id, hex_grid)
}
\arguments{
\item{full_grid_with_hexbin_id}{A data frame containing hexagon IDs, centroids, and standardized counts.}

\item{hex_grid}{A data frame containing all coordinates of hexagons.}
}
\value{
A data frame with hexagon information along with mapped polygon IDs.
}
\description{
This function maps polygon IDs to the corresponding hexagon coordinates in the full grid.
}
\examples{
nldr_df <- s_curve_noise_umap
num_bins <- 8
shape_val <- 2.031141
hexbin_data_object <- extract_hexbin_mean(nldr_df, num_bins, shape_val)
df_bin_centroids <- hexbin_data_object$hexdf_data
full_centroid_df <- generate_full_grid_centroids(df_bin_centroids)
hex_grid <- full_hex_grid(full_centroid_df)
full_grid_with_hexbin_id <- map_hexbin_id(full_centroid_df, df_bin_centroids)
map_polygon_id(full_grid_with_hexbin_id, hex_grid)

}
