% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{predict_2d_embeddings}
\alias{predict_2d_embeddings}
\title{Predict 2D Embeddings}
\usage{
predict_2d_embeddings(test_data, df_bin_centroids, df_bin, type_NLDR = "UMAP")
}
\arguments{
\item{test_data}{The test dataset containing high-dimensional features and an ID column.}

\item{df_bin_centroids}{Centroid coordinates of hexagonal bins in 2D space.}

\item{df_bin}{Centroid coordinates of hexagonal bins in high dimensions.}

\item{type_NLDR}{The type of non-linear dimensionality reduction (NLDR) used. Default is "UMAP".}
}
\value{
A data frame with predicted 2D embedding for each data point in the test dataset.
}
\description{
Given a test dataset, the centroid coordinates of hexagonal bins in  2D and high-dimensional space,
predict the 2D embeddings for each data point in the test dataset.
}
\examples{
model <- fit_high_d_model(training_data = s_curve_noise_training,
nldr_df_with_id = s_curve_noise_umap)
df_bin_centroids <- model$df_bin_centroids
df_bin <- model$df_bin
predict_2d_embeddings(test_data = s_curve_noise_test, df_bin_centroids = df_bin_centroids,
df_bin = df_bin, type_NLDR = "UMAP")

}
