% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{predict_hex_id}
\alias{predict_hex_id}
\title{Predict Hexagonal IDs}
\usage{
predict_hex_id(df_bin_centroids, nldr_df_test, x = "UMAP1", y = "UMAP2")
}
\arguments{
\item{df_bin_centroids}{The training dataset containing high-dimensional data with IDs.}

\item{nldr_df_test}{The non-linear dimensionality reductions that need to find the prediction.}

\item{x}{The name of the column that contains first embedding.}

\item{y}{The name of the column that contains second embedding.}
}
\value{
A data frame containing prediced hexID for 2D embedding data
}
\description{
This function predicts hexagonal IDs for a test set based on existing bin centroids.
}
\examples{
num_bins_x <- 4
shape_value <- 1.833091
hexbin_data_object <- extract_hexbin_mean(nldr_df = s_curve_noise_umap, num_bins_x,
shape_val = shape_value)
df_bin_centroids <- hexbin_data_object$hexdf_data
predict_hex_id(df_bin_centroids = df_bin_centroids, nldr_df_test = s_curve_noise_umap,
x = "UMAP1", y = "UMAP2")

}
