% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediction.R
\name{predict_hex_id}
\alias{predict_hex_id}
\title{Predict Hexagonal IDs}
\usage{
predict_hex_id(training_data, nldr_df, nldr_df_test, num_bins, shape_val)
}
\arguments{
\item{training_data}{The training dataset containing high-dimensional data with IDs.}

\item{nldr_df}{The non-linear dimensionality reduction data frame containing 2D coordinates of the training data.}

\item{nldr_df_test}{The predicted non-linear dimensionality reductions for the test set.}

\item{num_bins}{Number of bins along the x-axis for hexagon binning.}

\item{shape_val}{The value of the shape parameter for hexagon binning.}
}
\value{
A list containing the predicted data, hexagonal bin centroids, and averaged high-dimensional data.
}
\description{
This function predicts hexagonal IDs for a test set based on training data and non-linear dimensionality reduction (NLDR) coordinates.
}
\examples{
nldr_df <- s_curve_noise_umap
nldr_df_test <- s_curve_noise_umap_predict
training_data <- s_curve_noise_training
num_bins <- 8
shape_val <- 2.031141
hexbin_data_object <- extract_hexbin_mean(nldr_df, num_bins, shape_val)
df_bin_centroids <- hexbin_data_object$hexdf_data
predict_hex_id(training_data, nldr_df, nldr_df_test, num_bins, shape_val)

}
