% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2D.R
\name{remove_long_edges}
\alias{remove_long_edges}
\title{Remove Long Edges from a Triangular Mesh Plot}
\usage{
remove_long_edges(.data, benchmark_value, triangular_object, distance_col)
}
\arguments{
\item{.data}{The data frame containing the edge information.}

\item{benchmark_value}{The threshold value to determine long edges. Edges with a distance greater than or equal to this value will be removed.}

\item{triangular_object}{The triangular object containing the mesh information.}

\item{distance_col}{The column name in `.data` representing the distances.}
}
\value{
A ggplot object with the triangular mesh plot where long edges are removed.
}
\description{
This function removes long edges from a triangular mesh plot based on a benchmark value.
}
\examples{
nldr_df <- readRDS(paste0(here::here(), "/quollr/data-raw/s_curve_noise_umap.rds"))
num_bins <- 8
shape_val <- 2.031141
hexbin_data_object <- extract_hexbin_mean(nldr_df, num_bins, shape_val)
df_bin_centroids <- hexbin_data_object$hexdf_data
tr1_object <- triangulate_bin_centroids(df_bin_centroids, x, y)
tr_from_to_df <- generate_edge_info(triangular_object = tr1_object)
distance_df <- cal_2D_dist(tr_from_to_df)
remove_long_edges(distance_df, 0.6, tr1_object, "distance")

}
