% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scurve_model_obj}
\alias{scurve_model_obj}
\title{Object for S-curve dataset}
\format{
## `scurve_model_obj`
An object of five elements
\describe{
  \item{nldr_scaled_obj}{A tibble contains scaled first and second columns of NLDR
  data, and numeric vectors representing the limits of the original NLDR data.}
  \item{hb_obj}{A object that contains hexagonal binning information.}
  \item{model_highd}{A tibble with high-dimensional model.}
  \item{model_2d}{A tibble containing hexagonal bin centroids in 2-D}
  \item{trimesh_data}{A tibble that contains the edge information.}
}
}
\source{
This object is generated for illustrative purposes.
}
\usage{
data(scurve_model_obj)
}
\description{
The `scurve_model_obj` contains object of scaled umap embedding, x and y limits,
range of y; object of hexagonal binning information,
object of high-d model fitted to umap embedding for the training data,
the triangular object representing the triangulated bin centroids,
a tibble that contains the edge information, a tibble with
edge distance information.
}
\examples{
# Load the scurve_model_obj
data(scurve_model_obj)

}
\keyword{datasets}
