% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highd.R
\name{show_langevitour}
\alias{show_langevitour}
\title{Visualize the model overlaid on high-dimensional data}
\usage{
show_langevitour(point_data, edge_data)
}
\arguments{
\item{point_data}{A tibble that contains the high-dimensional data and model in high-dimensions.}

\item{edge_data}{A tibble that contains the wireframe data (from and to).}
}
\value{
A langevitour object with the model and the high-dimensional data.
}
\description{
This function generates a LangeviTour visualization based on different
conditions and input parameters.
}
\examples{
df_exe <- comb_data_model(highd_data = scurve, model_highd = scurve_model_obj$model_highd,
model_2d = scurve_model_obj$model_2d)
edge_data <- scurve_model_obj$trimesh_data
show_langevitour(point_data = df_exe, edge_data = edge_data)

}
