% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highD.R
\name{show_langevitour}
\alias{show_langevitour}
\title{Show LangeviTour Visualization}
\usage{
show_langevitour(
  df,
  df_b,
  df_b_with_center_data,
  benchmark_value = NA,
  distance_df,
  distance_col,
  min_points_threshold = NA
)
}
\arguments{
\item{df}{The original dataset.}

\item{df_b}{The summarized mean dataset.}

\item{df_b_with_center_data}{The dataset with hexbin centroids.}

\item{benchmark_value}{The benchmark value used to remove long edges (optional).}

\item{distance_df}{The distance dataframe.}

\item{distance_col}{The name of the distance column.}

\item{min_points_threshold}{The minimum number of points threshold for filtering bin centroids (optional).}
}
\description{
This function generates a LangeviTour visualization based on different conditions and input parameters.
}
\examples{
df <- tibble::tribble(
  ~x, ~y, ~hb_id,
  1, 2, 1,
  4, 5, 2,
  7, 8, 3
)
df_b <- tibble::tribble(
  ~x, ~y, ~hb_id,
  2, 3, 1,
  5, 6, 2,
  8, 9, 3
)
df_b_with_center_data <- tibble::tribble(
  ~x_val_center, ~y_val_center, ~hb_id,
  1.5, 2.5, 1,
  4.5, 5.5, 2,
  7.5, 8.5, 3
)
distance_df <- tibble::tribble(
  ~from, ~to, ~distance,
  1, 2, 0.5,
  2, 3, 0.7,
  1, 3, 0.2
)
show_langevitour(df, df_b, df_b_with_center_data, benchmark_value = 0.6, distance_df, distance_col = distance)

}
