% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nldr_data_link_plots.R
\name{show_link_plots}
\alias{show_link_plots}
\title{Visualize the model overlaid on high-dimensional data along with 2D wireframe model.}
\usage{
show_link_plots(point_df, edge_df)
}
\arguments{
\item{point_df}{A tibble that contains the high-dimensional data, non-linear dimension reductions
and model in high-dimensions.}

\item{edge_df}{A tibble that contains the wireframe data (from and to).}
}
\value{
A browsable HTML element.
}
\description{
This function generates a LangeviTour visualization based on different
conditions and input parameters with 2D wireframe.
}
\examples{
df_bin_centroids <- s_curve_obj$s_curve_umap_model_obj$df_bin_centroids
df_bin <- s_curve_obj$s_curve_umap_model_obj$df_bin
df_exe <- comb_all_data_model(highd_data = s_curve_noise_training,
nldr_data = s_curve_obj$s_curve_umap_scaled_obj$scaled_nldr,
model_highd = df_bin, model_2d = df_bin_centroids)
distance_df <- s_curve_obj$s_curve_umap_model_distance_df
benchmark <- find_lg_benchmark(distance_edges = distance_df,
distance_col = "distance")
distance_small_df <- distance_df |> dplyr::filter(distance < benchmark)
show_link_plots(point_df = df_exe, edge_df = distance_small_df)

}
