% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_2d.R
\name{tri_bin_centroids}
\alias{tri_bin_centroids}
\title{Triangulate bin centroids}
\usage{
tri_bin_centroids(hex_df, x, y)
}
\arguments{
\item{hex_df}{The tibble containing the bin centroids.}

\item{x}{The name of the column that contains x coordinates of bin centroids.}

\item{y}{The name of the column that contains y coordinates of bin centroids.}
}
\value{
A triangular object representing the triangulated bin centroids.
}
\description{
This function triangulates the bin centroids using the x and y coordinates
provided in the input data frame and returns the triangular object.
}
\examples{
df_bin_centroids <- s_curve_obj$s_curve_umap_model_obj$df_bin_centroids
tri_bin_centroids(hex_df = df_bin_centroids, x = "c_x", y = "c_y")

}
