% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2D.R
\name{triangulate_bin_centroids}
\alias{triangulate_bin_centroids}
\title{Triangulate Bin Centroids}
\usage{
triangulate_bin_centroids(.data, x, y)
}
\arguments{
\item{.data}{The data frame containing the bin centroids.}

\item{x}{The name of the column that contains x coordinates of centroids.}

\item{y}{The name of the column that contains y coordinates of centroids.}
}
\value{
A triangular object representing the triangulated bin centroids.
}
\description{
This function triangulates the bin centroids using the x and y coordinates provided in the input data frame and returns the triangular object.
}
\examples{
nldr_df <- s_curve_noise_umap
num_bins <- 7
shape_val <- 1.833091
hexbin_data_object <- extract_hexbin_mean(nldr_df, num_bins, shape_val)
df_bin_centroids <- hexbin_data_object$hexdf_data
triangulate_bin_centroids(df_bin_centroids, x, y)

}
