% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/2D.R
\name{triangulate_bin_centroids}
\alias{triangulate_bin_centroids}
\title{Triangulate Bin Centroids}
\usage{
triangulate_bin_centroids(.data, x, y)
}
\arguments{
\item{.data}{The data frame containing the bin centroids/ means.}

\item{x}{The name of the column that contains x coordinates of bin centroids/ means.}

\item{y}{The name of the column that contains y coordinates of bin centroids/ means.}
}
\value{
A triangular object representing the triangulated bin centroids/ means.
}
\description{
This function triangulates the bin centroids/ means using the x and y coordinates provided in the input data frame and returns the triangular object.
}
\examples{
num_bins_x <- 4
shape_value <- 1.833091
hexbin_data_object <- extract_hexbin_centroids(nldr_df = s_curve_noise_umap,
num_bins = num_bins_x, shape_val = shape_value)
df_bin_centroids <- hexbin_data_object$hexdf_data
triangulate_bin_centroids(df_bin_centroids, x, y)

}
